Struct aws_sdk_sagemaker::client::fluent_builders::CreateModelPackage

pub struct CreateModelPackage<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> CreateModelPackage<C>

pub async fn send(
self
) -> Result<CreateModelPackageOutput, SdkError<CreateModelPackageError>> where
C: SmithyConnector,

pub fn model_package_name(self, input: impl Into<String>) -> Self

The name of the model package. The name must have 1 to 63 characters. Valid characters are a-z, A-Z, 0-9, and - (hyphen).

This parameter is required for unversioned models. It is not applicable to versioned models.

pub fn set_model_package_name(self, input: Option<String>) -> Self

pub fn model_package_group_name(self, input: impl Into<String>) -> Self

The name of the model group that this model version belongs to.

This parameter is required for versioned models, and does not apply to unversioned models.

pub fn set_model_package_group_name(self, input: Option<String>) -> Self

pub fn model_package_description(self, input: impl Into<String>) -> Self

A description of the model package.

pub fn set_model_package_description(self, input: Option<String>) -> Self

pub fn inference_specification(self, input: InferenceSpecification) -> Self

Specifies details about inference jobs that can be run with models based on this model package, including the following:

The Amazon ECR paths of containers that contain the inference code and model artifacts.

The instance types that the model package supports for transform jobs and real-time endpoints used for inference.

The input and output content formats that the model package supports for inference.

pub fn set_inference_specification(
self,
input: Option<InferenceSpecification>
) -> Self

pub fn validation_specification(
self,
input: ModelPackageValidationSpecification
) -> Self

Specifies configurations for one or more transform jobs that Amazon SageMaker runs to test the model package.

pub fn set_validation_specification(
self,
input: Option<ModelPackageValidationSpecification>
) -> Self

pub fn source_algorithm_specification(
self,
input: SourceAlgorithmSpecification
) -> Self

Details about the algorithm that was used to create the model package.

pub fn set_source_algorithm_specification(
self,
input: Option<SourceAlgorithmSpecification>
) -> Self

pub fn certify_for_marketplace(self, input: bool) -> Self

Whether to certify the model package for listing on Amazon Web Services Marketplace.

This parameter is optional for unversioned models, and does not apply to versioned models.

pub fn set_certify_for_marketplace(self, input: Option<bool>) -> Self

pub fn tags(self, inp: impl Into<Tag>) -> Self

A list of key value pairs associated with the model. For more information, see Tagging Amazon Web Services resources in the Amazon Web Services General Reference Guide.

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn model_approval_status(self, input: ModelApprovalStatus) -> Self

Whether the model is approved for deployment.

This parameter is optional for versioned models, and does not apply to unversioned models.

For versioned models, the value of this parameter must be set to Approved to deploy the model.

pub fn set_model_approval_status(
self,
input: Option<ModelApprovalStatus>
) -> Self

pub fn metadata_properties(self, input: MetadataProperties) -> Self

Metadata properties of the tracking entity, trial, or trial component.

pub fn set_metadata_properties(self, input: Option<MetadataProperties>) -> Self

pub fn model_metrics(self, input: ModelMetrics) -> Self

A structure that contains model metrics reports.

pub fn set_model_metrics(self, input: Option<ModelMetrics>) -> Self

pub fn client_token(self, input: impl Into<String>) -> Self

A unique token that guarantees that the call to this API is idempotent.

pub fn set_client_token(self, input: Option<String>) -> Self