Struct aws_sdk_sagemaker::client::fluent_builders::CreateNotebookInstance

pub struct CreateNotebookInstance<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> CreateNotebookInstance<C>

pub async fn send(
self
) -> Result<CreateNotebookInstanceOutput, SdkError<CreateNotebookInstanceError>> where
C: SmithyConnector,

pub fn notebook_instance_name(self, input: impl Into<String>) -> Self

The name of the new notebook instance.

pub fn set_notebook_instance_name(self, input: Option<String>) -> Self

pub fn instance_type(self, input: InstanceType) -> Self

The type of ML compute instance to launch for the notebook instance.

pub fn set_instance_type(self, input: Option<InstanceType>) -> Self

pub fn subnet_id(self, input: impl Into<String>) -> Self

The ID of the subnet in a VPC to which you would like to have a connectivity from your ML compute instance.

pub fn set_subnet_id(self, input: Option<String>) -> Self

pub fn security_group_ids(self, inp: impl Into<String>) -> Self

The VPC security group IDs, in the form sg-xxxxxxxx. The security groups must be for the same VPC as specified in the subnet.

pub fn set_security_group_ids(self, input: Option<Vec<String>>) -> Self

pub fn role_arn(self, input: impl Into<String>) -> Self

When you send any requests to Amazon Web Services resources from the notebook instance, Amazon SageMaker assumes this role to perform tasks on your behalf. You must grant this role necessary permissions so Amazon SageMaker can perform these tasks. The policy must allow the Amazon SageMaker service principal (sagemaker.amazonaws.com) permissions to assume this role. For more information, see Amazon SageMaker Roles.

To be able to pass this role to Amazon SageMaker, the caller of this API must have the iam:PassRole permission.

pub fn set_role_arn(self, input: Option<String>) -> Self

pub fn kms_key_id(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of a Amazon Web Services Key Management Service key that Amazon SageMaker uses to encrypt data on the storage volume attached to your notebook instance. The KMS key you provide must be enabled. For information, see Enabling and Disabling Keys in the Amazon Web Services Key Management Service Developer Guide.

pub fn set_kms_key_id(self, input: Option<String>) -> Self

pub fn tags(self, inp: impl Into<Tag>) -> Self

An array of key-value pairs. You can use tags to categorize your Amazon Web Services resources in different ways, for example, by purpose, owner, or environment. For more information, see Tagging Amazon Web Services Resources.

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn lifecycle_config_name(self, input: impl Into<String>) -> Self

The name of a lifecycle configuration to associate with the notebook instance. For information about lifestyle configurations, see Step 2.1: (Optional) Customize a Notebook Instance.

pub fn set_lifecycle_config_name(self, input: Option<String>) -> Self

pub fn direct_internet_access(self, input: DirectInternetAccess) -> Self

Sets whether Amazon SageMaker provides internet access to the notebook instance. If you set this to Disabled this notebook instance is able to access resources only in your VPC, and is not be able to connect to Amazon SageMaker training and endpoint services unless you configure a NAT Gateway in your VPC.

For more information, see Notebook Instances Are Internet-Enabled by Default. You can set the value of this parameter to Disabled only if you set a value for the SubnetId parameter.

pub fn set_direct_internet_access(
self,
input: Option<DirectInternetAccess>
) -> Self

pub fn volume_size_in_gb(self, input: i32) -> Self

The size, in GB, of the ML storage volume to attach to the notebook instance. The default value is 5 GB.

pub fn set_volume_size_in_gb(self, input: Option<i32>) -> Self

pub fn accelerator_types(
self,
inp: impl Into<NotebookInstanceAcceleratorType>
) -> Self

A list of Elastic Inference (EI) instance types to associate with this notebook instance. Currently, only one instance type can be associated with a notebook instance. For more information, see Using Elastic Inference in Amazon SageMaker.

pub fn set_accelerator_types(
self,
input: Option<Vec<NotebookInstanceAcceleratorType>>
) -> Self

pub fn default_code_repository(self, input: impl Into<String>) -> Self

A Git repository to associate with the notebook instance as its default code repository. This can be either the name of a Git repository stored as a resource in your account, or the URL of a Git repository in Amazon Web Services CodeCommit or in any other Git repository. When you open a notebook instance, it opens in the directory that contains this repository. For more information, see Associating Git Repositories with Amazon SageMaker Notebook Instances.

pub fn set_default_code_repository(self, input: Option<String>) -> Self

pub fn additional_code_repositories(self, inp: impl Into<String>) -> Self

An array of up to three Git repositories to associate with the notebook instance. These can be either the names of Git repositories stored as resources in your account, or the URL of Git repositories in Amazon Web Services CodeCommit or in any other Git repository. These repositories are cloned at the same level as the default repository of your notebook instance. For more information, see Associating Git Repositories with Amazon SageMaker Notebook Instances.

pub fn set_additional_code_repositories(
self,
input: Option<Vec<String>>
) -> Self

pub fn root_access(self, input: RootAccess) -> Self

Whether root access is enabled or disabled for users of the notebook instance. The default value is Enabled.

Lifecycle configurations need root access to be able to set up a notebook instance. Because of this, lifecycle configurations associated with a notebook instance always run with root access even if you disable root access for users.

pub fn set_root_access(self, input: Option<RootAccess>) -> Self