Struct aws_sdk_sagemaker::client::fluent_builders::CreateTrainingJob

pub struct CreateTrainingJob<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> CreateTrainingJob<C>

pub async fn send(
self
) -> Result<CreateTrainingJobOutput, SdkError<CreateTrainingJobError>> where
C: SmithyConnector,

pub fn training_job_name(self, input: impl Into<String>) -> Self

The name of the training job. The name must be unique within an Amazon Web Services Region in an Amazon Web Services account.

pub fn set_training_job_name(self, input: Option<String>) -> Self

pub fn hyper_parameters(
self,
k: impl Into<String>,
v: impl Into<String>
) -> Self

Algorithm-specific parameters that influence the quality of the model. You set hyperparameters before you start the learning process. For a list of hyperparameters for each training algorithm provided by Amazon SageMaker, see Algorithms.

You can specify a maximum of 100 hyperparameters. Each hyperparameter is a key-value pair. Each key and value is limited to 256 characters, as specified by the Length Constraint.

pub fn set_hyper_parameters(
self,
input: Option<HashMap<String, String>>
) -> Self

pub fn algorithm_specification(self, input: AlgorithmSpecification) -> Self

The registry path of the Docker image that contains the training algorithm and algorithm-specific metadata, including the input mode. For more information about algorithms provided by Amazon SageMaker, see Algorithms. For information about providing your own algorithms, see Using Your Own Algorithms with Amazon SageMaker.

pub fn set_algorithm_specification(
self,
input: Option<AlgorithmSpecification>
) -> Self

pub fn role_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of an IAM role that Amazon SageMaker can assume to perform tasks on your behalf.

During model training, Amazon SageMaker needs your permission to read input data from an S3 bucket, download a Docker image that contains training code, write model artifacts to an S3 bucket, write logs to Amazon CloudWatch Logs, and publish metrics to Amazon CloudWatch. You grant permissions for all of these tasks to an IAM role. For more information, see Amazon SageMaker Roles.

To be able to pass this role to Amazon SageMaker, the caller of this API must have the iam:PassRole permission.

pub fn set_role_arn(self, input: Option<String>) -> Self

pub fn input_data_config(self, inp: impl Into<Channel>) -> Self

An array of Channel objects. Each channel is a named input source. InputDataConfig describes the input data and its location.

Algorithms can accept input data from one or more channels. For example, an algorithm might have two channels of input data, training_data and validation_data. The configuration for each channel provides the S3, EFS, or FSx location where the input data is stored. It also provides information about the stored data: the MIME type, compression method, and whether the data is wrapped in RecordIO format.

Depending on the input mode that the algorithm supports, Amazon SageMaker either copies input data files from an S3 bucket to a local directory in the Docker container, or makes it available as input streams. For example, if you specify an EFS location, input data files will be made available as input streams. They do not need to be downloaded.

pub fn set_input_data_config(self, input: Option<Vec<Channel>>) -> Self

pub fn output_data_config(self, input: OutputDataConfig) -> Self

Specifies the path to the S3 location where you want to store model artifacts. Amazon SageMaker creates subfolders for the artifacts.

pub fn set_output_data_config(self, input: Option<OutputDataConfig>) -> Self

pub fn resource_config(self, input: ResourceConfig) -> Self

The resources, including the ML compute instances and ML storage volumes, to use for model training.

ML storage volumes store model artifacts and incremental states. Training algorithms might also use ML storage volumes for scratch space. If you want Amazon SageMaker to use the ML storage volume to store the training data, choose File as the TrainingInputMode in the algorithm specification. For distributed training algorithms, specify an instance count greater than 1.

pub fn set_resource_config(self, input: Option<ResourceConfig>) -> Self

pub fn vpc_config(self, input: VpcConfig) -> Self

A VpcConfig object that specifies the VPC that you want your training job to connect to. Control access to and from your training container by configuring the VPC. For more information, see Protect Training Jobs by Using an Amazon Virtual Private Cloud.

pub fn set_vpc_config(self, input: Option<VpcConfig>) -> Self

pub fn stopping_condition(self, input: StoppingCondition) -> Self

Specifies a limit to how long a model training job can run. It also specifies how long a managed Spot training job has to complete. When the job reaches the time limit, Amazon SageMaker ends the training job. Use this API to cap model training costs.

To stop a job, Amazon SageMaker sends the algorithm the SIGTERM signal, which delays job termination for 120 seconds. Algorithms can use this 120-second window to save the model artifacts, so the results of training are not lost.

pub fn set_stopping_condition(self, input: Option<StoppingCondition>) -> Self

pub fn tags(self, inp: impl Into<Tag>) -> Self

An array of key-value pairs. You can use tags to categorize your Amazon Web Services resources in different ways, for example, by purpose, owner, or environment. For more information, see Tagging Amazon Web Services Resources.

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn enable_network_isolation(self, input: bool) -> Self

Isolates the training container. No inbound or outbound network calls can be made, except for calls between peers within a training cluster for distributed training. If you enable network isolation for training jobs that are configured to use a VPC, Amazon SageMaker downloads and uploads customer data and model artifacts through the specified VPC, but the training container does not have network access.

pub fn set_enable_network_isolation(self, input: Option<bool>) -> Self

pub fn enable_inter_container_traffic_encryption(self, input: bool) -> Self

To encrypt all communications between ML compute instances in distributed training, choose True. Encryption provides greater security for distributed training, but training might take longer. How long it takes depends on the amount of communication between compute instances, especially if you use a deep learning algorithm in distributed training. For more information, see Protect Communications Between ML Compute Instances in a Distributed Training Job.

pub fn set_enable_inter_container_traffic_encryption(
self,
input: Option<bool>
) -> Self

pub fn enable_managed_spot_training(self, input: bool) -> Self

To train models using managed spot training, choose True. Managed spot training provides a fully managed and scalable infrastructure for training machine learning models. this option is useful when training jobs can be interrupted and when there is flexibility when the training job is run.

The complete and intermediate results of jobs are stored in an Amazon S3 bucket, and can be used as a starting point to train models incrementally. Amazon SageMaker provides metrics and logs in CloudWatch. They can be used to see when managed spot training jobs are running, interrupted, resumed, or completed.

pub fn set_enable_managed_spot_training(self, input: Option<bool>) -> Self

pub fn checkpoint_config(self, input: CheckpointConfig) -> Self

Contains information about the output location for managed spot training checkpoint data.

pub fn set_checkpoint_config(self, input: Option<CheckpointConfig>) -> Self

pub fn debug_hook_config(self, input: DebugHookConfig) -> Self

Configuration information for the Debugger hook parameters, metric and tensor collections, and storage paths. To learn more about how to configure the DebugHookConfig parameter, see Use the SageMaker and Debugger Configuration API Operations to Create, Update, and Debug Your Training Job.

pub fn set_debug_hook_config(self, input: Option<DebugHookConfig>) -> Self

pub fn debug_rule_configurations(
self,
inp: impl Into<DebugRuleConfiguration>
) -> Self

Configuration information for Debugger rules for debugging output tensors.

pub fn set_debug_rule_configurations(
self,
input: Option<Vec<DebugRuleConfiguration>>
) -> Self

pub fn tensor_board_output_config(self, input: TensorBoardOutputConfig) -> Self

Configuration of storage locations for the Debugger TensorBoard output data.

pub fn set_tensor_board_output_config(
self,
input: Option<TensorBoardOutputConfig>
) -> Self

pub fn experiment_config(self, input: ExperimentConfig) -> Self

Associates a SageMaker job as a trial component with an experiment and trial. Specified when you call the following APIs:

CreateProcessingJob

CreateTrainingJob

CreateTransformJob

pub fn set_experiment_config(self, input: Option<ExperimentConfig>) -> Self

pub fn profiler_config(self, input: ProfilerConfig) -> Self

Configuration information for Debugger system monitoring, framework profiling, and storage paths.

pub fn set_profiler_config(self, input: Option<ProfilerConfig>) -> Self

pub fn profiler_rule_configurations(
self,
inp: impl Into<ProfilerRuleConfiguration>
) -> Self

Configuration information for Debugger rules for profiling system and framework metrics.

pub fn set_profiler_rule_configurations(
self,
input: Option<Vec<ProfilerRuleConfiguration>>
) -> Self

pub fn environment(self, k: impl Into<String>, v: impl Into<String>) -> Self

The environment variables to set in the Docker container.

pub fn set_environment(self, input: Option<HashMap<String, String>>) -> Self

pub fn retry_strategy(self, input: RetryStrategy) -> Self

The number of times to retry the job when the job fails due to an InternalServerError.

pub fn set_retry_strategy(self, input: Option<RetryStrategy>) -> Self