Struct aws_sdk_sagemaker::client::fluent_builders::CreateTransformJob

pub struct CreateTransformJob<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> CreateTransformJob<C>

pub async fn send(
self
) -> Result<CreateTransformJobOutput, SdkError<CreateTransformJobError>> where
C: SmithyConnector,

pub fn transform_job_name(self, input: impl Into<String>) -> Self

The name of the transform job. The name must be unique within an Amazon Web Services Region in an Amazon Web Services account.

pub fn set_transform_job_name(self, input: Option<String>) -> Self

pub fn model_name(self, input: impl Into<String>) -> Self

The name of the model that you want to use for the transform job. ModelName must be the name of an existing Amazon SageMaker model within an Amazon Web Services Region in an Amazon Web Services account.

pub fn set_model_name(self, input: Option<String>) -> Self

pub fn max_concurrent_transforms(self, input: i32) -> Self

The maximum number of parallel requests that can be sent to each instance in a transform job. If MaxConcurrentTransforms is set to 0 or left unset, Amazon SageMaker checks the optional execution-parameters to determine the settings for your chosen algorithm. If the execution-parameters endpoint is not enabled, the default value is 1. For more information on execution-parameters, see How Containers Serve Requests. For built-in algorithms, you don't need to set a value for MaxConcurrentTransforms.

pub fn set_max_concurrent_transforms(self, input: Option<i32>) -> Self

pub fn model_client_config(self, input: ModelClientConfig) -> Self

Configures the timeout and maximum number of retries for processing a transform job invocation.

pub fn set_model_client_config(self, input: Option<ModelClientConfig>) -> Self

pub fn max_payload_in_mb(self, input: i32) -> Self

The maximum allowed size of the payload, in MB. A payload is the data portion of a record (without metadata). The value in MaxPayloadInMB must be greater than, or equal to, the size of a single record. To estimate the size of a record in MB, divide the size of your dataset by the number of records. To ensure that the records fit within the maximum payload size, we recommend using a slightly larger value. The default value is 6 MB.

For cases where the payload might be arbitrarily large and is transmitted using HTTP chunked encoding, set the value to 0. This feature works only in supported algorithms. Currently, Amazon SageMaker built-in algorithms do not support HTTP chunked encoding.

pub fn set_max_payload_in_mb(self, input: Option<i32>) -> Self

pub fn batch_strategy(self, input: BatchStrategy) -> Self

Specifies the number of records to include in a mini-batch for an HTTP inference request. A record is a single unit of input data that inference can be made on. For example, a single line in a CSV file is a record.

To enable the batch strategy, you must set the SplitType property to Line, RecordIO, or TFRecord.

To use only one record when making an HTTP invocation request to a container, set BatchStrategy to SingleRecord and SplitType to Line.

To fit as many records in a mini-batch as can fit within the MaxPayloadInMB limit, set BatchStrategy to MultiRecord and SplitType to Line.

pub fn set_batch_strategy(self, input: Option<BatchStrategy>) -> Self

pub fn environment(self, k: impl Into<String>, v: impl Into<String>) -> Self

The environment variables to set in the Docker container. We support up to 16 key and values entries in the map.

pub fn set_environment(self, input: Option<HashMap<String, String>>) -> Self

pub fn transform_input(self, input: TransformInput) -> Self

Describes the input source and the way the transform job consumes it.

pub fn set_transform_input(self, input: Option<TransformInput>) -> Self

pub fn transform_output(self, input: TransformOutput) -> Self

Describes the results of the transform job.

pub fn set_transform_output(self, input: Option<TransformOutput>) -> Self

pub fn transform_resources(self, input: TransformResources) -> Self

Describes the resources, including ML instance types and ML instance count, to use for the transform job.

pub fn set_transform_resources(self, input: Option<TransformResources>) -> Self

pub fn data_processing(self, input: DataProcessing) -> Self

The data structure used to specify the data to be used for inference in a batch transform job and to associate the data that is relevant to the prediction results in the output. The input filter provided allows you to exclude input data that is not needed for inference in a batch transform job. The output filter provided allows you to include input data relevant to interpreting the predictions in the output from the job. For more information, see Associate Prediction Results with their Corresponding Input Records.

pub fn set_data_processing(self, input: Option<DataProcessing>) -> Self

pub fn tags(self, inp: impl Into<Tag>) -> Self

(Optional) An array of key-value pairs. For more information, see Using Cost Allocation Tags in the Amazon Web Services Billing and Cost Management User Guide.

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn experiment_config(self, input: ExperimentConfig) -> Self

Associates a SageMaker job as a trial component with an experiment and trial. Specified when you call the following APIs:

CreateProcessingJob

CreateTrainingJob

CreateTransformJob

pub fn set_experiment_config(self, input: Option<ExperimentConfig>) -> Self