Struct aws_sdk_sagemaker::client::fluent_builders::CreateTrialComponent

pub struct CreateTrialComponent<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> CreateTrialComponent<C>

pub async fn send(
self
) -> Result<CreateTrialComponentOutput, SdkError<CreateTrialComponentError>> where
C: SmithyConnector,

pub fn trial_component_name(self, input: impl Into<String>) -> Self

The name of the component. The name must be unique in your Amazon Web Services account and is not case-sensitive.

pub fn set_trial_component_name(self, input: Option<String>) -> Self

pub fn display_name(self, input: impl Into<String>) -> Self

The name of the component as displayed. The name doesn't need to be unique. If DisplayName isn't specified, TrialComponentName is displayed.

pub fn set_display_name(self, input: Option<String>) -> Self

pub fn status(self, input: TrialComponentStatus) -> Self

The status of the component. States include:

InProgress

Completed

Failed

pub fn set_status(self, input: Option<TrialComponentStatus>) -> Self

pub fn start_time(self, input: Instant) -> Self

When the component started.

pub fn set_start_time(self, input: Option<Instant>) -> Self

pub fn end_time(self, input: Instant) -> Self

When the component ended.

pub fn set_end_time(self, input: Option<Instant>) -> Self

pub fn parameters(
self,
k: impl Into<String>,
v: impl Into<TrialComponentParameterValue>
) -> Self

The hyperparameters for the component.

pub fn set_parameters(
self,
input: Option<HashMap<String, TrialComponentParameterValue>>
) -> Self

pub fn input_artifacts(
self,
k: impl Into<String>,
v: impl Into<TrialComponentArtifact>
) -> Self

The input artifacts for the component. Examples of input artifacts are datasets, algorithms, hyperparameters, source code, and instance types.

pub fn set_input_artifacts(
self,
input: Option<HashMap<String, TrialComponentArtifact>>
) -> Self

pub fn output_artifacts(
self,
k: impl Into<String>,
v: impl Into<TrialComponentArtifact>
) -> Self

The output artifacts for the component. Examples of output artifacts are metrics, snapshots, logs, and images.

pub fn set_output_artifacts(
self,
input: Option<HashMap<String, TrialComponentArtifact>>
) -> Self

pub fn metadata_properties(self, input: MetadataProperties) -> Self

Metadata properties of the tracking entity, trial, or trial component.

pub fn set_metadata_properties(self, input: Option<MetadataProperties>) -> Self

pub fn tags(self, inp: impl Into<Tag>) -> Self

A list of tags to associate with the component. You can use Search API to search on the tags.

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self