Struct aws_sdk_sagemaker::client::fluent_builders::ListMonitoringSchedules

pub struct ListMonitoringSchedules<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> ListMonitoringSchedules<C>

pub async fn send(
self
) -> Result<ListMonitoringSchedulesOutput, SdkError<ListMonitoringSchedulesError>> where
C: SmithyConnector,

pub fn endpoint_name(self, input: impl Into<String>) -> Self

Name of a specific endpoint to fetch schedules for.

pub fn set_endpoint_name(self, input: Option<String>) -> Self

pub fn sort_by(self, input: MonitoringScheduleSortKey) -> Self

Whether to sort results by Status, CreationTime, ScheduledTime field. The default is CreationTime.

pub fn set_sort_by(self, input: Option<MonitoringScheduleSortKey>) -> Self

pub fn sort_order(self, input: SortOrder) -> Self

Whether to sort the results in Ascending or Descending order. The default is Descending.

pub fn set_sort_order(self, input: Option<SortOrder>) -> Self

pub fn next_token(self, input: impl Into<String>) -> Self

The token returned if the response is truncated. To retrieve the next set of job executions, use it in the next request.

pub fn set_next_token(self, input: Option<String>) -> Self

pub fn max_results(self, input: i32) -> Self

The maximum number of jobs to return in the response. The default value is 10.

pub fn set_max_results(self, input: Option<i32>) -> Self

pub fn name_contains(self, input: impl Into<String>) -> Self

Filter for monitoring schedules whose name contains a specified string.

pub fn set_name_contains(self, input: Option<String>) -> Self

pub fn creation_time_before(self, input: Instant) -> Self

A filter that returns only monitoring schedules created before a specified time.

pub fn set_creation_time_before(self, input: Option<Instant>) -> Self

pub fn creation_time_after(self, input: Instant) -> Self

A filter that returns only monitoring schedules created after a specified time.

pub fn set_creation_time_after(self, input: Option<Instant>) -> Self

pub fn last_modified_time_before(self, input: Instant) -> Self

A filter that returns only monitoring schedules modified before a specified time.

pub fn set_last_modified_time_before(self, input: Option<Instant>) -> Self

pub fn last_modified_time_after(self, input: Instant) -> Self

A filter that returns only monitoring schedules modified after a specified time.

pub fn set_last_modified_time_after(self, input: Option<Instant>) -> Self

pub fn status_equals(self, input: ScheduleStatus) -> Self

A filter that returns only monitoring schedules modified before a specified time.

pub fn set_status_equals(self, input: Option<ScheduleStatus>) -> Self

pub fn monitoring_job_definition_name(self, input: impl Into<String>) -> Self

Gets a list of the monitoring schedules for the specified monitoring job definition.

pub fn set_monitoring_job_definition_name(self, input: Option<String>) -> Self

pub fn monitoring_type_equals(self, input: MonitoringType) -> Self

A filter that returns only the monitoring schedules for the specified monitoring type.

pub fn set_monitoring_type_equals(self, input: Option<MonitoringType>) -> Self