Struct aws_sdk_sagemaker::client::fluent_builders::UpdateNotebookInstance

pub struct UpdateNotebookInstance<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> UpdateNotebookInstance<C>

pub async fn send(
self
) -> Result<UpdateNotebookInstanceOutput, SdkError<UpdateNotebookInstanceError>> where
C: SmithyConnector,

pub fn notebook_instance_name(self, input: impl Into<String>) -> Self

The name of the notebook instance to update.

pub fn set_notebook_instance_name(self, input: Option<String>) -> Self

pub fn instance_type(self, input: InstanceType) -> Self

The Amazon ML compute instance type.

pub fn set_instance_type(self, input: Option<InstanceType>) -> Self

pub fn role_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the IAM role that Amazon SageMaker can assume to access the notebook instance. For more information, see Amazon SageMaker Roles.

To be able to pass this role to Amazon SageMaker, the caller of this API must have the iam:PassRole permission.

pub fn set_role_arn(self, input: Option<String>) -> Self

pub fn lifecycle_config_name(self, input: impl Into<String>) -> Self

The name of a lifecycle configuration to associate with the notebook instance. For information about lifestyle configurations, see Step 2.1: (Optional) Customize a Notebook Instance.

pub fn set_lifecycle_config_name(self, input: Option<String>) -> Self

pub fn disassociate_lifecycle_config(self, input: bool) -> Self

Set to true to remove the notebook instance lifecycle configuration currently associated with the notebook instance. This operation is idempotent. If you specify a lifecycle configuration that is not associated with the notebook instance when you call this method, it does not throw an error.

pub fn set_disassociate_lifecycle_config(self, input: Option<bool>) -> Self

pub fn volume_size_in_gb(self, input: i32) -> Self

The size, in GB, of the ML storage volume to attach to the notebook instance. The default value is 5 GB. ML storage volumes are encrypted, so Amazon SageMaker can't determine the amount of available free space on the volume. Because of this, you can increase the volume size when you update a notebook instance, but you can't decrease the volume size. If you want to decrease the size of the ML storage volume in use, create a new notebook instance with the desired size.

pub fn set_volume_size_in_gb(self, input: Option<i32>) -> Self

pub fn default_code_repository(self, input: impl Into<String>) -> Self

The Git repository to associate with the notebook instance as its default code repository. This can be either the name of a Git repository stored as a resource in your account, or the URL of a Git repository in Amazon Web Services CodeCommit or in any other Git repository. When you open a notebook instance, it opens in the directory that contains this repository. For more information, see Associating Git Repositories with Amazon SageMaker Notebook Instances.

pub fn set_default_code_repository(self, input: Option<String>) -> Self

pub fn additional_code_repositories(self, inp: impl Into<String>) -> Self

An array of up to three Git repositories to associate with the notebook instance. These can be either the names of Git repositories stored as resources in your account, or the URL of Git repositories in Amazon Web Services CodeCommit or in any other Git repository. These repositories are cloned at the same level as the default repository of your notebook instance. For more information, see Associating Git Repositories with Amazon SageMaker Notebook Instances.

pub fn set_additional_code_repositories(
self,
input: Option<Vec<String>>
) -> Self

pub fn accelerator_types(
self,
inp: impl Into<NotebookInstanceAcceleratorType>
) -> Self

A list of the Elastic Inference (EI) instance types to associate with this notebook instance. Currently only one EI instance type can be associated with a notebook instance. For more information, see Using Elastic Inference in Amazon SageMaker.

pub fn set_accelerator_types(
self,
input: Option<Vec<NotebookInstanceAcceleratorType>>
) -> Self

pub fn disassociate_accelerator_types(self, input: bool) -> Self

A list of the Elastic Inference (EI) instance types to remove from this notebook instance. This operation is idempotent. If you specify an accelerator type that is not associated with the notebook instance when you call this method, it does not throw an error.

pub fn set_disassociate_accelerator_types(self, input: Option<bool>) -> Self

pub fn disassociate_default_code_repository(self, input: bool) -> Self

The name or URL of the default Git repository to remove from this notebook instance. This operation is idempotent. If you specify a Git repository that is not associated with the notebook instance when you call this method, it does not throw an error.

pub fn set_disassociate_default_code_repository(
self,
input: Option<bool>
) -> Self

pub fn disassociate_additional_code_repositories(self, input: bool) -> Self

A list of names or URLs of the default Git repositories to remove from this notebook instance. This operation is idempotent. If you specify a Git repository that is not associated with the notebook instance when you call this method, it does not throw an error.

pub fn set_disassociate_additional_code_repositories(
self,
input: Option<bool>
) -> Self

pub fn root_access(self, input: RootAccess) -> Self

Whether root access is enabled or disabled for users of the notebook instance. The default value is Enabled.

If you set this to Disabled, users don't have root access on the notebook instance, but lifecycle configuration scripts still run with root permissions.

pub fn set_root_access(self, input: Option<RootAccess>) -> Self