Struct aws_sdk_sagemaker::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for SageMaker.

This client allows ergonomic access to a SageMaker-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn add_association(&self) -> AddAssociation<C, M, R>

pub fn add_tags(&self) -> AddTags<C, M, R>

pub fn associate_trial_component(&self) -> AssociateTrialComponent<C, M, R>

pub fn create_action(&self) -> CreateAction<C, M, R>

pub fn create_algorithm(&self) -> CreateAlgorithm<C, M, R>

pub fn create_app(&self) -> CreateApp<C, M, R>

pub fn create_app_image_config(&self) -> CreateAppImageConfig<C, M, R>

pub fn create_artifact(&self) -> CreateArtifact<C, M, R>

pub fn create_auto_ml_job(&self) -> CreateAutoMLJob<C, M, R>

pub fn create_code_repository(&self) -> CreateCodeRepository<C, M, R>

pub fn create_compilation_job(&self) -> CreateCompilationJob<C, M, R>

pub fn create_context(&self) -> CreateContext<C, M, R>

pub fn create_data_quality_job_definition(
&self
) -> CreateDataQualityJobDefinition<C, M, R>

pub fn create_device_fleet(&self) -> CreateDeviceFleet<C, M, R>

pub fn create_domain(&self) -> CreateDomain<C, M, R>

pub fn create_edge_packaging_job(&self) -> CreateEdgePackagingJob<C, M, R>

pub fn create_endpoint(&self) -> CreateEndpoint<C, M, R>

pub fn create_endpoint_config(&self) -> CreateEndpointConfig<C, M, R>

pub fn create_experiment(&self) -> CreateExperiment<C, M, R>

pub fn create_feature_group(&self) -> CreateFeatureGroup<C, M, R>

pub fn create_flow_definition(&self) -> CreateFlowDefinition<C, M, R>

pub fn create_human_task_ui(&self) -> CreateHumanTaskUi<C, M, R>

pub fn create_hyper_parameter_tuning_job(
&self
) -> CreateHyperParameterTuningJob<C, M, R>

pub fn create_image(&self) -> CreateImage<C, M, R>

pub fn create_image_version(&self) -> CreateImageVersion<C, M, R>

pub fn create_labeling_job(&self) -> CreateLabelingJob<C, M, R>

pub fn create_model(&self) -> CreateModel<C, M, R>

pub fn create_model_bias_job_definition(
&self
) -> CreateModelBiasJobDefinition<C, M, R>

pub fn create_model_explainability_job_definition(
&self
) -> CreateModelExplainabilityJobDefinition<C, M, R>

pub fn create_model_package(&self) -> CreateModelPackage<C, M, R>

pub fn create_model_package_group(&self) -> CreateModelPackageGroup<C, M, R>

pub fn create_model_quality_job_definition(
&self
) -> CreateModelQualityJobDefinition<C, M, R>

pub fn create_monitoring_schedule(&self) -> CreateMonitoringSchedule<C, M, R>

pub fn create_notebook_instance(&self) -> CreateNotebookInstance<C, M, R>

pub fn create_notebook_instance_lifecycle_config(
&self
) -> CreateNotebookInstanceLifecycleConfig<C, M, R>

pub fn create_pipeline(&self) -> CreatePipeline<C, M, R>

pub fn create_presigned_domain_url(&self) -> CreatePresignedDomainUrl<C, M, R>

pub fn create_presigned_notebook_instance_url(
&self
) -> CreatePresignedNotebookInstanceUrl<C, M, R>

pub fn create_processing_job(&self) -> CreateProcessingJob<C, M, R>

pub fn create_project(&self) -> CreateProject<C, M, R>

pub fn create_studio_lifecycle_config(
&self
) -> CreateStudioLifecycleConfig<C, M, R>

pub fn create_training_job(&self) -> CreateTrainingJob<C, M, R>

pub fn create_transform_job(&self) -> CreateTransformJob<C, M, R>

pub fn create_trial(&self) -> CreateTrial<C, M, R>

pub fn create_trial_component(&self) -> CreateTrialComponent<C, M, R>

pub fn create_user_profile(&self) -> CreateUserProfile<C, M, R>

pub fn create_workforce(&self) -> CreateWorkforce<C, M, R>

pub fn create_workteam(&self) -> CreateWorkteam<C, M, R>

pub fn delete_action(&self) -> DeleteAction<C, M, R>

pub fn delete_algorithm(&self) -> DeleteAlgorithm<C, M, R>

pub fn delete_app(&self) -> DeleteApp<C, M, R>

pub fn delete_app_image_config(&self) -> DeleteAppImageConfig<C, M, R>

pub fn delete_artifact(&self) -> DeleteArtifact<C, M, R>

pub fn delete_association(&self) -> DeleteAssociation<C, M, R>

pub fn delete_code_repository(&self) -> DeleteCodeRepository<C, M, R>

pub fn delete_context(&self) -> DeleteContext<C, M, R>

pub fn delete_data_quality_job_definition(
&self
) -> DeleteDataQualityJobDefinition<C, M, R>

pub fn delete_device_fleet(&self) -> DeleteDeviceFleet<C, M, R>

pub fn delete_domain(&self) -> DeleteDomain<C, M, R>

pub fn delete_endpoint(&self) -> DeleteEndpoint<C, M, R>

pub fn delete_endpoint_config(&self) -> DeleteEndpointConfig<C, M, R>

pub fn delete_experiment(&self) -> DeleteExperiment<C, M, R>

pub fn delete_feature_group(&self) -> DeleteFeatureGroup<C, M, R>

pub fn delete_flow_definition(&self) -> DeleteFlowDefinition<C, M, R>

pub fn delete_human_task_ui(&self) -> DeleteHumanTaskUi<C, M, R>

pub fn delete_image(&self) -> DeleteImage<C, M, R>

pub fn delete_image_version(&self) -> DeleteImageVersion<C, M, R>

pub fn delete_model(&self) -> DeleteModel<C, M, R>

pub fn delete_model_bias_job_definition(
&self
) -> DeleteModelBiasJobDefinition<C, M, R>

pub fn delete_model_explainability_job_definition(
&self
) -> DeleteModelExplainabilityJobDefinition<C, M, R>

pub fn delete_model_package(&self) -> DeleteModelPackage<C, M, R>

pub fn delete_model_package_group(&self) -> DeleteModelPackageGroup<C, M, R>

pub fn delete_model_package_group_policy(
&self
) -> DeleteModelPackageGroupPolicy<C, M, R>

pub fn delete_model_quality_job_definition(
&self
) -> DeleteModelQualityJobDefinition<C, M, R>

pub fn delete_monitoring_schedule(&self) -> DeleteMonitoringSchedule<C, M, R>

pub fn delete_notebook_instance(&self) -> DeleteNotebookInstance<C, M, R>

pub fn delete_notebook_instance_lifecycle_config(
&self
) -> DeleteNotebookInstanceLifecycleConfig<C, M, R>

pub fn delete_pipeline(&self) -> DeletePipeline<C, M, R>

pub fn delete_project(&self) -> DeleteProject<C, M, R>

pub fn delete_studio_lifecycle_config(
&self
) -> DeleteStudioLifecycleConfig<C, M, R>

pub fn delete_tags(&self) -> DeleteTags<C, M, R>

pub fn delete_trial(&self) -> DeleteTrial<C, M, R>

pub fn delete_trial_component(&self) -> DeleteTrialComponent<C, M, R>

pub fn delete_user_profile(&self) -> DeleteUserProfile<C, M, R>

pub fn delete_workforce(&self) -> DeleteWorkforce<C, M, R>

pub fn delete_workteam(&self) -> DeleteWorkteam<C, M, R>

pub fn deregister_devices(&self) -> DeregisterDevices<C, M, R>

pub fn describe_action(&self) -> DescribeAction<C, M, R>

pub fn describe_algorithm(&self) -> DescribeAlgorithm<C, M, R>

pub fn describe_app(&self) -> DescribeApp<C, M, R>

pub fn describe_app_image_config(&self) -> DescribeAppImageConfig<C, M, R>

pub fn describe_artifact(&self) -> DescribeArtifact<C, M, R>

pub fn describe_auto_ml_job(&self) -> DescribeAutoMLJob<C, M, R>

pub fn describe_code_repository(&self) -> DescribeCodeRepository<C, M, R>

pub fn describe_compilation_job(&self) -> DescribeCompilationJob<C, M, R>

pub fn describe_context(&self) -> DescribeContext<C, M, R>

pub fn describe_data_quality_job_definition(
&self
) -> DescribeDataQualityJobDefinition<C, M, R>

pub fn describe_device(&self) -> DescribeDevice<C, M, R>

pub fn describe_device_fleet(&self) -> DescribeDeviceFleet<C, M, R>

pub fn describe_domain(&self) -> DescribeDomain<C, M, R>

pub fn describe_edge_packaging_job(&self) -> DescribeEdgePackagingJob<C, M, R>

pub fn describe_endpoint(&self) -> DescribeEndpoint<C, M, R>

pub fn describe_endpoint_config(&self) -> DescribeEndpointConfig<C, M, R>

pub fn describe_experiment(&self) -> DescribeExperiment<C, M, R>

pub fn describe_feature_group(&self) -> DescribeFeatureGroup<C, M, R>

pub fn describe_flow_definition(&self) -> DescribeFlowDefinition<C, M, R>

pub fn describe_human_task_ui(&self) -> DescribeHumanTaskUi<C, M, R>

pub fn describe_hyper_parameter_tuning_job(
&self
) -> DescribeHyperParameterTuningJob<C, M, R>

pub fn describe_image(&self) -> DescribeImage<C, M, R>

pub fn describe_image_version(&self) -> DescribeImageVersion<C, M, R>

pub fn describe_labeling_job(&self) -> DescribeLabelingJob<C, M, R>

pub fn describe_model(&self) -> DescribeModel<C, M, R>

pub fn describe_model_bias_job_definition(
&self
) -> DescribeModelBiasJobDefinition<C, M, R>

pub fn describe_model_explainability_job_definition(
&self
) -> DescribeModelExplainabilityJobDefinition<C, M, R>

pub fn describe_model_package(&self) -> DescribeModelPackage<C, M, R>

pub fn describe_model_package_group(&self) -> DescribeModelPackageGroup<C, M, R>

pub fn describe_model_quality_job_definition(
&self
) -> DescribeModelQualityJobDefinition<C, M, R>

pub fn describe_monitoring_schedule(
&self
) -> DescribeMonitoringSchedule<C, M, R>

pub fn describe_notebook_instance(&self) -> DescribeNotebookInstance<C, M, R>

pub fn describe_notebook_instance_lifecycle_config(
&self
) -> DescribeNotebookInstanceLifecycleConfig<C, M, R>

pub fn describe_pipeline(&self) -> DescribePipeline<C, M, R>

pub fn describe_pipeline_definition_for_execution(
&self
) -> DescribePipelineDefinitionForExecution<C, M, R>

pub fn describe_pipeline_execution(&self) -> DescribePipelineExecution<C, M, R>

pub fn describe_processing_job(&self) -> DescribeProcessingJob<C, M, R>

pub fn describe_project(&self) -> DescribeProject<C, M, R>

pub fn describe_studio_lifecycle_config(
&self
) -> DescribeStudioLifecycleConfig<C, M, R>

pub fn describe_subscribed_workteam(
&self
) -> DescribeSubscribedWorkteam<C, M, R>

pub fn describe_training_job(&self) -> DescribeTrainingJob<C, M, R>

pub fn describe_transform_job(&self) -> DescribeTransformJob<C, M, R>

pub fn describe_trial(&self) -> DescribeTrial<C, M, R>

pub fn describe_trial_component(&self) -> DescribeTrialComponent<C, M, R>

pub fn describe_user_profile(&self) -> DescribeUserProfile<C, M, R>

pub fn describe_workforce(&self) -> DescribeWorkforce<C, M, R>

pub fn describe_workteam(&self) -> DescribeWorkteam<C, M, R>

pub fn disable_sagemaker_servicecatalog_portfolio(
&self
) -> DisableSagemakerServicecatalogPortfolio<C, M, R>

pub fn disassociate_trial_component(
&self
) -> DisassociateTrialComponent<C, M, R>

pub fn enable_sagemaker_servicecatalog_portfolio(
&self
) -> EnableSagemakerServicecatalogPortfolio<C, M, R>

pub fn get_device_fleet_report(&self) -> GetDeviceFleetReport<C, M, R>

pub fn get_model_package_group_policy(
&self
) -> GetModelPackageGroupPolicy<C, M, R>

pub fn get_sagemaker_servicecatalog_portfolio_status(
&self
) -> GetSagemakerServicecatalogPortfolioStatus<C, M, R>

pub fn get_search_suggestions(&self) -> GetSearchSuggestions<C, M, R>

pub fn list_actions(&self) -> ListActions<C, M, R>

pub fn list_algorithms(&self) -> ListAlgorithms<C, M, R>

pub fn list_app_image_configs(&self) -> ListAppImageConfigs<C, M, R>

pub fn list_apps(&self) -> ListApps<C, M, R>

pub fn list_artifacts(&self) -> ListArtifacts<C, M, R>

pub fn list_associations(&self) -> ListAssociations<C, M, R>

pub fn list_auto_ml_jobs(&self) -> ListAutoMLJobs<C, M, R>

pub fn list_candidates_for_auto_ml_job(
&self
) -> ListCandidatesForAutoMLJob<C, M, R>

pub fn list_code_repositories(&self) -> ListCodeRepositories<C, M, R>

pub fn list_compilation_jobs(&self) -> ListCompilationJobs<C, M, R>

pub fn list_contexts(&self) -> ListContexts<C, M, R>

pub fn list_data_quality_job_definitions(
&self
) -> ListDataQualityJobDefinitions<C, M, R>

pub fn list_device_fleets(&self) -> ListDeviceFleets<C, M, R>

pub fn list_devices(&self) -> ListDevices<C, M, R>

pub fn list_domains(&self) -> ListDomains<C, M, R>

pub fn list_edge_packaging_jobs(&self) -> ListEdgePackagingJobs<C, M, R>

pub fn list_endpoint_configs(&self) -> ListEndpointConfigs<C, M, R>

pub fn list_endpoints(&self) -> ListEndpoints<C, M, R>

pub fn list_experiments(&self) -> ListExperiments<C, M, R>

pub fn list_feature_groups(&self) -> ListFeatureGroups<C, M, R>

pub fn list_flow_definitions(&self) -> ListFlowDefinitions<C, M, R>

pub fn list_human_task_uis(&self) -> ListHumanTaskUis<C, M, R>

pub fn list_hyper_parameter_tuning_jobs(
&self
) -> ListHyperParameterTuningJobs<C, M, R>

pub fn list_images(&self) -> ListImages<C, M, R>

pub fn list_image_versions(&self) -> ListImageVersions<C, M, R>

pub fn list_labeling_jobs(&self) -> ListLabelingJobs<C, M, R>

pub fn list_labeling_jobs_for_workteam(
&self
) -> ListLabelingJobsForWorkteam<C, M, R>

pub fn list_model_bias_job_definitions(
&self
) -> ListModelBiasJobDefinitions<C, M, R>

pub fn list_model_explainability_job_definitions(
&self
) -> ListModelExplainabilityJobDefinitions<C, M, R>

pub fn list_model_package_groups(&self) -> ListModelPackageGroups<C, M, R>

pub fn list_model_packages(&self) -> ListModelPackages<C, M, R>

pub fn list_model_quality_job_definitions(
&self
) -> ListModelQualityJobDefinitions<C, M, R>

pub fn list_models(&self) -> ListModels<C, M, R>

pub fn list_monitoring_executions(&self) -> ListMonitoringExecutions<C, M, R>

pub fn list_monitoring_schedules(&self) -> ListMonitoringSchedules<C, M, R>

pub fn list_notebook_instance_lifecycle_configs(
&self
) -> ListNotebookInstanceLifecycleConfigs<C, M, R>

pub fn list_notebook_instances(&self) -> ListNotebookInstances<C, M, R>

pub fn list_pipeline_executions(&self) -> ListPipelineExecutions<C, M, R>

pub fn list_pipeline_execution_steps(
&self
) -> ListPipelineExecutionSteps<C, M, R>

pub fn list_pipeline_parameters_for_execution(
&self
) -> ListPipelineParametersForExecution<C, M, R>

pub fn list_pipelines(&self) -> ListPipelines<C, M, R>

pub fn list_processing_jobs(&self) -> ListProcessingJobs<C, M, R>

pub fn list_projects(&self) -> ListProjects<C, M, R>

pub fn list_studio_lifecycle_configs(
&self
) -> ListStudioLifecycleConfigs<C, M, R>

pub fn list_subscribed_workteams(&self) -> ListSubscribedWorkteams<C, M, R>

pub fn list_tags(&self) -> ListTags<C, M, R>

pub fn list_training_jobs(&self) -> ListTrainingJobs<C, M, R>

pub fn list_training_jobs_for_hyper_parameter_tuning_job(
&self
) -> ListTrainingJobsForHyperParameterTuningJob<C, M, R>

pub fn list_transform_jobs(&self) -> ListTransformJobs<C, M, R>

pub fn list_trial_components(&self) -> ListTrialComponents<C, M, R>

pub fn list_trials(&self) -> ListTrials<C, M, R>

pub fn list_user_profiles(&self) -> ListUserProfiles<C, M, R>

pub fn list_workforces(&self) -> ListWorkforces<C, M, R>

pub fn list_workteams(&self) -> ListWorkteams<C, M, R>

pub fn put_model_package_group_policy(
&self
) -> PutModelPackageGroupPolicy<C, M, R>

pub fn register_devices(&self) -> RegisterDevices<C, M, R>

pub fn render_ui_template(&self) -> RenderUiTemplate<C, M, R>

pub fn search(&self) -> Search<C, M, R>

pub fn send_pipeline_execution_step_failure(
&self
) -> SendPipelineExecutionStepFailure<C, M, R>

pub fn send_pipeline_execution_step_success(
&self
) -> SendPipelineExecutionStepSuccess<C, M, R>

pub fn start_monitoring_schedule(&self) -> StartMonitoringSchedule<C, M, R>

pub fn start_notebook_instance(&self) -> StartNotebookInstance<C, M, R>

pub fn start_pipeline_execution(&self) -> StartPipelineExecution<C, M, R>

pub fn stop_auto_ml_job(&self) -> StopAutoMLJob<C, M, R>

pub fn stop_compilation_job(&self) -> StopCompilationJob<C, M, R>

pub fn stop_edge_packaging_job(&self) -> StopEdgePackagingJob<C, M, R>

pub fn stop_hyper_parameter_tuning_job(
&self
) -> StopHyperParameterTuningJob<C, M, R>

pub fn stop_labeling_job(&self) -> StopLabelingJob<C, M, R>

pub fn stop_monitoring_schedule(&self) -> StopMonitoringSchedule<C, M, R>

pub fn stop_notebook_instance(&self) -> StopNotebookInstance<C, M, R>

pub fn stop_pipeline_execution(&self) -> StopPipelineExecution<C, M, R>

pub fn stop_processing_job(&self) -> StopProcessingJob<C, M, R>

pub fn stop_training_job(&self) -> StopTrainingJob<C, M, R>

pub fn stop_transform_job(&self) -> StopTransformJob<C, M, R>

pub fn update_action(&self) -> UpdateAction<C, M, R>

pub fn update_app_image_config(&self) -> UpdateAppImageConfig<C, M, R>

pub fn update_artifact(&self) -> UpdateArtifact<C, M, R>

pub fn update_code_repository(&self) -> UpdateCodeRepository<C, M, R>

pub fn update_context(&self) -> UpdateContext<C, M, R>

pub fn update_device_fleet(&self) -> UpdateDeviceFleet<C, M, R>

pub fn update_devices(&self) -> UpdateDevices<C, M, R>

pub fn update_domain(&self) -> UpdateDomain<C, M, R>

pub fn update_endpoint(&self) -> UpdateEndpoint<C, M, R>

pub fn update_endpoint_weights_and_capacities(
&self
) -> UpdateEndpointWeightsAndCapacities<C, M, R>

pub fn update_experiment(&self) -> UpdateExperiment<C, M, R>

pub fn update_image(&self) -> UpdateImage<C, M, R>

pub fn update_model_package(&self) -> UpdateModelPackage<C, M, R>

pub fn update_monitoring_schedule(&self) -> UpdateMonitoringSchedule<C, M, R>

pub fn update_notebook_instance(&self) -> UpdateNotebookInstance<C, M, R>

pub fn update_notebook_instance_lifecycle_config(
&self
) -> UpdateNotebookInstanceLifecycleConfig<C, M, R>

pub fn update_pipeline(&self) -> UpdatePipeline<C, M, R>

pub fn update_pipeline_execution(&self) -> UpdatePipelineExecution<C, M, R>

pub fn update_training_job(&self) -> UpdateTrainingJob<C, M, R>

pub fn update_trial(&self) -> UpdateTrial<C, M, R>

pub fn update_trial_component(&self) -> UpdateTrialComponent<C, M, R>

pub fn update_user_profile(&self) -> UpdateUserProfile<C, M, R>

pub fn update_workforce(&self) -> UpdateWorkforce<C, M, R>

pub fn update_workteam(&self) -> UpdateWorkteam<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self