Struct aws_sdk_sagemaker::input::create_processing_job_input::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for CreateProcessingJobInput

Implementations

impl Builder

pub fn processing_inputs(self, input: impl Into<ProcessingInput>) -> Self

pub fn set_processing_inputs(self, input: Option<Vec<ProcessingInput>>) -> Self

pub fn processing_output_config(self, input: ProcessingOutputConfig) -> Self

Output configuration for the processing job.

pub fn set_processing_output_config(
self,
input: Option<ProcessingOutputConfig>
) -> Self

pub fn processing_job_name(self, input: impl Into<String>) -> Self

The name of the processing job. The name must be unique within an Amazon Web Services Region in the Amazon Web Services account.

pub fn set_processing_job_name(self, input: Option<String>) -> Self

pub fn processing_resources(self, input: ProcessingResources) -> Self

Identifies the resources, ML compute instances, and ML storage volumes to deploy for a processing job. In distributed training, you specify more than one instance.

pub fn set_processing_resources(
self,
input: Option<ProcessingResources>
) -> Self

pub fn stopping_condition(self, input: ProcessingStoppingCondition) -> Self

The time limit for how long the processing job is allowed to run.

pub fn set_stopping_condition(
self,
input: Option<ProcessingStoppingCondition>
) -> Self

pub fn app_specification(self, input: AppSpecification) -> Self

Configures the processing job to run a specified Docker container image.

pub fn set_app_specification(self, input: Option<AppSpecification>) -> Self

pub fn environment(self, k: impl Into<String>, v: impl Into<String>) -> Self

pub fn set_environment(self, input: Option<HashMap<String, String>>) -> Self

pub fn network_config(self, input: NetworkConfig) -> Self

Networking options for a processing job, such as whether to allow inbound and outbound network calls to and from processing containers, and the VPC subnets and security groups to use for VPC-enabled processing jobs.

pub fn set_network_config(self, input: Option<NetworkConfig>) -> Self

pub fn role_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of an IAM role that Amazon SageMaker can assume to perform tasks on your behalf.

pub fn set_role_arn(self, input: Option<String>) -> Self

pub fn tags(self, input: impl Into<Tag>) -> Self

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn experiment_config(self, input: ExperimentConfig) -> Self

Associates a SageMaker job as a trial component with an experiment and trial. Specified when you call the following APIs:

CreateProcessingJob

CreateTrainingJob

CreateTransformJob

pub fn set_experiment_config(self, input: Option<ExperimentConfig>) -> Self

pub fn build(self) -> Result<CreateProcessingJobInput, BuildError>

Consumes the builder and constructs a CreateProcessingJobInput