Struct aws_sdk_sagemaker::input::CreateTrainingJobInput

#[non_exhaustive]
pub struct CreateTrainingJobInput {

    pub training_job_name: Option<String>,
    pub hyper_parameters: Option<HashMap<String, String>>,
    pub algorithm_specification: Option<AlgorithmSpecification>,
    pub role_arn: Option<String>,
    pub input_data_config: Option<Vec<Channel>>,
    pub output_data_config: Option<OutputDataConfig>,
    pub resource_config: Option<ResourceConfig>,
    pub vpc_config: Option<VpcConfig>,
    pub stopping_condition: Option<StoppingCondition>,
    pub tags: Option<Vec<Tag>>,
    pub enable_network_isolation: bool,
    pub enable_inter_container_traffic_encryption: bool,
    pub enable_managed_spot_training: bool,
    pub checkpoint_config: Option<CheckpointConfig>,
    pub debug_hook_config: Option<DebugHookConfig>,
    pub debug_rule_configurations: Option<Vec<DebugRuleConfiguration>>,
    pub tensor_board_output_config: Option<TensorBoardOutputConfig>,
    pub experiment_config: Option<ExperimentConfig>,
    pub profiler_config: Option<ProfilerConfig>,
    pub profiler_rule_configurations: Option<Vec<ProfilerRuleConfiguration>>,
    pub environment: Option<HashMap<String, String>>,
    pub retry_strategy: Option<RetryStrategy>,

}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

training_job_name: Option<String>

The name of the training job. The name must be unique within an Amazon Web Services Region in an Amazon Web Services account.

hyper_parameters: Option<HashMap<String, String>>

Algorithm-specific parameters that influence the quality of the model. You set hyperparameters before you start the learning process. For a list of hyperparameters for each training algorithm provided by Amazon SageMaker, see Algorithms.

You can specify a maximum of 100 hyperparameters. Each hyperparameter is a key-value pair. Each key and value is limited to 256 characters, as specified by the Length Constraint.

algorithm_specification: Option<AlgorithmSpecification>

The registry path of the Docker image that contains the training algorithm and algorithm-specific metadata, including the input mode. For more information about algorithms provided by Amazon SageMaker, see Algorithms. For information about providing your own algorithms, see Using Your Own Algorithms with Amazon SageMaker.

role_arn: Option<String>

The Amazon Resource Name (ARN) of an IAM role that Amazon SageMaker can assume to perform tasks on your behalf.

During model training, Amazon SageMaker needs your permission to read input data from an S3 bucket, download a Docker image that contains training code, write model artifacts to an S3 bucket, write logs to Amazon CloudWatch Logs, and publish metrics to Amazon CloudWatch. You grant permissions for all of these tasks to an IAM role. For more information, see Amazon SageMaker Roles.

To be able to pass this role to Amazon SageMaker, the caller of this API must have the iam:PassRole permission.

input_data_config: Option<Vec<Channel>>

An array of Channel objects. Each channel is a named input source. InputDataConfig describes the input data and its location.

Algorithms can accept input data from one or more channels. For example, an algorithm might have two channels of input data, training_data and validation_data. The configuration for each channel provides the S3, EFS, or FSx location where the input data is stored. It also provides information about the stored data: the MIME type, compression method, and whether the data is wrapped in RecordIO format.

Depending on the input mode that the algorithm supports, Amazon SageMaker either copies input data files from an S3 bucket to a local directory in the Docker container, or makes it available as input streams. For example, if you specify an EFS location, input data files will be made available as input streams. They do not need to be downloaded.

output_data_config: Option<OutputDataConfig>

Specifies the path to the S3 location where you want to store model artifacts. Amazon SageMaker creates subfolders for the artifacts.

resource_config: Option<ResourceConfig>

The resources, including the ML compute instances and ML storage volumes, to use for model training.

ML storage volumes store model artifacts and incremental states. Training algorithms might also use ML storage volumes for scratch space. If you want Amazon SageMaker to use the ML storage volume to store the training data, choose File as the TrainingInputMode in the algorithm specification. For distributed training algorithms, specify an instance count greater than 1.

vpc_config: Option<VpcConfig>

A VpcConfig object that specifies the VPC that you want your training job to connect to. Control access to and from your training container by configuring the VPC. For more information, see Protect Training Jobs by Using an Amazon Virtual Private Cloud.

stopping_condition: Option<StoppingCondition>

Specifies a limit to how long a model training job can run. It also specifies how long a managed Spot training job has to complete. When the job reaches the time limit, Amazon SageMaker ends the training job. Use this API to cap model training costs.

To stop a job, Amazon SageMaker sends the algorithm the SIGTERM signal, which delays job termination for 120 seconds. Algorithms can use this 120-second window to save the model artifacts, so the results of training are not lost.

tags: Option<Vec<Tag>>

An array of key-value pairs. You can use tags to categorize your Amazon Web Services resources in different ways, for example, by purpose, owner, or environment. For more information, see Tagging Amazon Web Services Resources.

enable_network_isolation: bool

Isolates the training container. No inbound or outbound network calls can be made, except for calls between peers within a training cluster for distributed training. If you enable network isolation for training jobs that are configured to use a VPC, Amazon SageMaker downloads and uploads customer data and model artifacts through the specified VPC, but the training container does not have network access.

enable_inter_container_traffic_encryption: bool

To encrypt all communications between ML compute instances in distributed training, choose True. Encryption provides greater security for distributed training, but training might take longer. How long it takes depends on the amount of communication between compute instances, especially if you use a deep learning algorithm in distributed training. For more information, see Protect Communications Between ML Compute Instances in a Distributed Training Job.

enable_managed_spot_training: bool

To train models using managed spot training, choose True. Managed spot training provides a fully managed and scalable infrastructure for training machine learning models. this option is useful when training jobs can be interrupted and when there is flexibility when the training job is run.

The complete and intermediate results of jobs are stored in an Amazon S3 bucket, and can be used as a starting point to train models incrementally. Amazon SageMaker provides metrics and logs in CloudWatch. They can be used to see when managed spot training jobs are running, interrupted, resumed, or completed.

checkpoint_config: Option<CheckpointConfig>

Contains information about the output location for managed spot training checkpoint data.

debug_hook_config: Option<DebugHookConfig>

Configuration information for the Debugger hook parameters, metric and tensor collections, and storage paths. To learn more about how to configure the DebugHookConfig parameter, see Use the SageMaker and Debugger Configuration API Operations to Create, Update, and Debug Your Training Job.

debug_rule_configurations: Option<Vec<DebugRuleConfiguration>>

Configuration information for Debugger rules for debugging output tensors.

tensor_board_output_config: Option<TensorBoardOutputConfig>

Configuration of storage locations for the Debugger TensorBoard output data.

experiment_config: Option<ExperimentConfig>

Associates a SageMaker job as a trial component with an experiment and trial. Specified when you call the following APIs:

CreateProcessingJob

CreateTrainingJob

CreateTransformJob

profiler_config: Option<ProfilerConfig>

Configuration information for Debugger system monitoring, framework profiling, and storage paths.

profiler_rule_configurations: Option<Vec<ProfilerRuleConfiguration>>

Configuration information for Debugger rules for profiling system and framework metrics.

environment: Option<HashMap<String, String>>

The environment variables to set in the Docker container.

retry_strategy: Option<RetryStrategy>

The number of times to retry the job when the job fails due to an InternalServerError.

Implementations

impl CreateTrainingJobInput

pub fn make_operation(
&self,
_config: &Config
) -> Result<Operation<CreateTrainingJob, AwsErrorRetryPolicy>, BuildError>

Consumes the builder and constructs an Operation<CreateTrainingJob>

pub fn builder() -> Builder

Creates a new builder-style object to manufacture CreateTrainingJobInput