Module aws_sdk_sagemaker::model

Modules

action_source

See ActionSource

action_summary

See ActionSummary

agent_version

See AgentVersion

alarm

See Alarm

algorithm_specification

See AlgorithmSpecification

algorithm_status_details

See AlgorithmStatusDetails

algorithm_status_item

See AlgorithmStatusItem

algorithm_summary

See AlgorithmSummary

algorithm_validation_profile

See AlgorithmValidationProfile

algorithm_validation_specification

See AlgorithmValidationSpecification

annotation_consolidation_config

See AnnotationConsolidationConfig

app_details

See AppDetails

app_image_config_details

See AppImageConfigDetails

app_specification

See AppSpecification

artifact_source

See ArtifactSource

artifact_source_type

See ArtifactSourceType

artifact_summary

See ArtifactSummary

association_summary

See AssociationSummary

athena_dataset_definition

See AthenaDatasetDefinition

auto_ml_candidate

See AutoMlCandidate

auto_ml_candidate_step

See AutoMlCandidateStep

auto_ml_channel

See AutoMlChannel

auto_ml_container_definition

See AutoMlContainerDefinition

auto_ml_data_source

See AutoMlDataSource

auto_ml_job_artifacts

See AutoMlJobArtifacts

auto_ml_job_completion_criteria

See AutoMlJobCompletionCriteria

auto_ml_job_config

See AutoMlJobConfig

auto_ml_job_objective

See AutoMlJobObjective

auto_ml_job_summary

See AutoMlJobSummary

auto_ml_output_data_config

See AutoMlOutputDataConfig

auto_ml_partial_failure_reason

See AutoMlPartialFailureReason

auto_ml_security_config

See AutoMlSecurityConfig

auto_mls3_data_source

See AutoMls3DataSource

auto_rollback_config

See AutoRollbackConfig

bias

See Bias

blue_green_update_policy

See BlueGreenUpdatePolicy

cache_hit_result

See CacheHitResult

callback_step_metadata

See CallbackStepMetadata

candidate_artifact_locations

See CandidateArtifactLocations

candidate_properties

See CandidateProperties

capacity_size

See CapacitySize

capture_content_type_header

See CaptureContentTypeHeader

capture_option

See CaptureOption

categorical_parameter_range

See CategoricalParameterRange

categorical_parameter_range_specification

See CategoricalParameterRangeSpecification

channel

See Channel

channel_specification

See ChannelSpecification

checkpoint_config

See CheckpointConfig

code_repository_summary

See CodeRepositorySummary

cognito_config

See CognitoConfig

cognito_member_definition

See CognitoMemberDefinition

collection_configuration

See CollectionConfiguration

compilation_job_summary

See CompilationJobSummary

condition_step_metadata

See ConditionStepMetadata

container_definition

See ContainerDefinition

context_source

See ContextSource

context_summary

See ContextSummary

continuous_parameter_range

See ContinuousParameterRange

continuous_parameter_range_specification

See ContinuousParameterRangeSpecification

custom_image

See CustomImage

data_capture_config

See DataCaptureConfig

data_capture_config_summary

See DataCaptureConfigSummary

data_catalog_config

See DataCatalogConfig

data_processing

See DataProcessing

data_quality_app_specification

See DataQualityAppSpecification

data_quality_baseline_config

See DataQualityBaselineConfig

data_quality_job_input

See DataQualityJobInput

data_source

See DataSource

dataset_definition

See DatasetDefinition

debug_hook_config

See DebugHookConfig

debug_rule_configuration

See DebugRuleConfiguration

debug_rule_evaluation_status

See DebugRuleEvaluationStatus

deployed_image

See DeployedImage

deployment_config

See DeploymentConfig

desired_weight_and_capacity

See DesiredWeightAndCapacity

device

See Device

device_fleet_summary

See DeviceFleetSummary

device_stats

See DeviceStats

device_summary

See DeviceSummary

domain_details

See DomainDetails

edge_model

See EdgeModel

edge_model_stat

See EdgeModelStat

edge_model_summary

See EdgeModelSummary

edge_output_config

See EdgeOutputConfig

edge_packaging_job_summary

See EdgePackagingJobSummary

edge_preset_deployment_output

See EdgePresetDeploymentOutput

endpoint

See Endpoint

endpoint_config_summary

See EndpointConfigSummary

endpoint_input

See EndpointInput

endpoint_summary

See EndpointSummary

experiment

See Experiment

experiment_config

See ExperimentConfig

experiment_source

See ExperimentSource

experiment_summary

See ExperimentSummary

explainability

See Explainability

feature_definition

See FeatureDefinition

feature_group

See FeatureGroup

feature_group_summary

See FeatureGroupSummary

file_system_config

See FileSystemConfig

file_system_data_source

See FileSystemDataSource

filter

See Filter

final_auto_ml_job_objective_metric

See FinalAutoMlJobObjectiveMetric

final_hyper_parameter_tuning_job_objective_metric

See FinalHyperParameterTuningJobObjectiveMetric

flow_definition_output_config

See FlowDefinitionOutputConfig

flow_definition_summary

See FlowDefinitionSummary

git_config

See GitConfig

git_config_for_update

See GitConfigForUpdate

human_loop_activation_conditions_config

See HumanLoopActivationConditionsConfig

human_loop_activation_config

See HumanLoopActivationConfig

human_loop_config

See HumanLoopConfig

human_loop_request_source

See HumanLoopRequestSource

human_task_config

See HumanTaskConfig

human_task_ui_summary

See HumanTaskUiSummary

hyper_parameter_algorithm_specification

See HyperParameterAlgorithmSpecification

hyper_parameter_specification

See HyperParameterSpecification

hyper_parameter_training_job_definition

See HyperParameterTrainingJobDefinition

hyper_parameter_training_job_summary

See HyperParameterTrainingJobSummary

hyper_parameter_tuning_job_config

See HyperParameterTuningJobConfig

hyper_parameter_tuning_job_objective

See HyperParameterTuningJobObjective

hyper_parameter_tuning_job_summary

See HyperParameterTuningJobSummary

hyper_parameter_tuning_job_warm_start_config

See HyperParameterTuningJobWarmStartConfig

image

See Image

image_config

See ImageConfig

image_version

See ImageVersion

inference_execution_config

See InferenceExecutionConfig

inference_specification

See InferenceSpecification

input_config

See InputConfig

integer_parameter_range

See IntegerParameterRange

integer_parameter_range_specification

See IntegerParameterRangeSpecification

jupyter_server_app_settings

See JupyterServerAppSettings

kernel_gateway_app_settings

See KernelGatewayAppSettings

kernel_gateway_image_config

See KernelGatewayImageConfig

kernel_spec

See KernelSpec

label_counters

See LabelCounters

label_counters_for_workteam

See LabelCountersForWorkteam

labeling_job_algorithms_config

See LabelingJobAlgorithmsConfig

labeling_job_data_attributes

See LabelingJobDataAttributes

labeling_job_data_source

See LabelingJobDataSource

labeling_job_for_workteam_summary

See LabelingJobForWorkteamSummary

labeling_job_input_config

See LabelingJobInputConfig

labeling_job_output

See LabelingJobOutput

labeling_job_output_config

See LabelingJobOutputConfig

labeling_job_resource_config

See LabelingJobResourceConfig

labeling_job_s3_data_source

See LabelingJobS3DataSource

labeling_job_sns_data_source

See LabelingJobSnsDataSource

labeling_job_stopping_conditions

See LabelingJobStoppingConditions

labeling_job_summary

See LabelingJobSummary

member_definition

See MemberDefinition

metadata_properties

See MetadataProperties

metric_data

See MetricData

metric_definition

See MetricDefinition

metrics_source

See MetricsSource

model_artifacts

See ModelArtifacts

model_bias_app_specification

See ModelBiasAppSpecification

model_bias_baseline_config

See ModelBiasBaselineConfig

model_bias_job_input

See ModelBiasJobInput

model_client_config

See ModelClientConfig

model_data_quality

See ModelDataQuality

model_deploy_config

See ModelDeployConfig

model_deploy_result

See ModelDeployResult

model_digests

See ModelDigests

model_explainability_app_specification

See ModelExplainabilityAppSpecification

model_explainability_baseline_config

See ModelExplainabilityBaselineConfig

model_explainability_job_input

See ModelExplainabilityJobInput

model_metrics

See ModelMetrics

model_package

See ModelPackage

model_package_container_definition

See ModelPackageContainerDefinition

model_package_group

See ModelPackageGroup

model_package_group_summary

See ModelPackageGroupSummary

model_package_status_details

See ModelPackageStatusDetails

model_package_status_item

See ModelPackageStatusItem

model_package_summary

See ModelPackageSummary

model_package_validation_profile

See ModelPackageValidationProfile

model_package_validation_specification

See ModelPackageValidationSpecification

model_quality

See ModelQuality

model_quality_app_specification

See ModelQualityAppSpecification

model_quality_baseline_config

See ModelQualityBaselineConfig

model_quality_job_input

See ModelQualityJobInput

model_step_metadata

See ModelStepMetadata

model_summary

See ModelSummary

monitoring_app_specification

See MonitoringAppSpecification

monitoring_baseline_config

See MonitoringBaselineConfig

monitoring_cluster_config

See MonitoringClusterConfig

monitoring_constraints_resource

See MonitoringConstraintsResource

monitoring_execution_summary

See MonitoringExecutionSummary

monitoring_ground_truth_s3_input

See MonitoringGroundTruthS3Input

monitoring_input

See MonitoringInput

monitoring_job_definition

See MonitoringJobDefinition

monitoring_job_definition_summary

See MonitoringJobDefinitionSummary

monitoring_network_config

See MonitoringNetworkConfig

monitoring_output

See MonitoringOutput

monitoring_output_config

See MonitoringOutputConfig

monitoring_resources

See MonitoringResources

monitoring_s3_output

See MonitoringS3Output

monitoring_schedule

See MonitoringSchedule

monitoring_schedule_config

See MonitoringScheduleConfig

monitoring_schedule_summary

See MonitoringScheduleSummary

monitoring_statistics_resource

See MonitoringStatisticsResource

monitoring_stopping_condition

See MonitoringStoppingCondition

multi_model_config

See MultiModelConfig

neo_vpc_config

See NeoVpcConfig

nested_filters

See NestedFilters

network_config

See NetworkConfig

notebook_instance_lifecycle_config_summary

See NotebookInstanceLifecycleConfigSummary

notebook_instance_lifecycle_hook

See NotebookInstanceLifecycleHook

notebook_instance_summary

See NotebookInstanceSummary

notification_configuration

See NotificationConfiguration

objective_status_counters

See ObjectiveStatusCounters

offline_store_config

See OfflineStoreConfig

offline_store_status

See OfflineStoreStatus

oidc_config

See OidcConfig

oidc_config_for_response

See OidcConfigForResponse

oidc_member_definition

See OidcMemberDefinition

online_store_config

See OnlineStoreConfig

online_store_security_config

See OnlineStoreSecurityConfig

output_config

See OutputConfig

output_data_config

See OutputDataConfig

output_parameter

See OutputParameter

parameter

See Parameter

parameter_range

See ParameterRange

parameter_ranges

See ParameterRanges

parent

See Parent

parent_hyper_parameter_tuning_job

See ParentHyperParameterTuningJob

pipeline

See Pipeline

pipeline_execution

See PipelineExecution

pipeline_execution_step

See PipelineExecutionStep

pipeline_execution_step_metadata

See PipelineExecutionStepMetadata

pipeline_execution_summary

See PipelineExecutionSummary

pipeline_experiment_config

See PipelineExperimentConfig

pipeline_summary

See PipelineSummary

processing_cluster_config

See ProcessingClusterConfig

processing_feature_store_output

See ProcessingFeatureStoreOutput

processing_input

See ProcessingInput

processing_job

See ProcessingJob

processing_job_step_metadata

See ProcessingJobStepMetadata

processing_job_summary

See ProcessingJobSummary

processing_output

See ProcessingOutput

processing_output_config

See ProcessingOutputConfig

processing_resources

See ProcessingResources

processing_s3_input

See ProcessingS3Input

processing_s3_output

See ProcessingS3Output

processing_stopping_condition

See ProcessingStoppingCondition

production_variant

See ProductionVariant

production_variant_core_dump_config

See ProductionVariantCoreDumpConfig

production_variant_summary

See ProductionVariantSummary

profiler_config

See ProfilerConfig

profiler_config_for_update

See ProfilerConfigForUpdate

profiler_rule_configuration

See ProfilerRuleConfiguration

profiler_rule_evaluation_status

See ProfilerRuleEvaluationStatus

project_summary

See ProjectSummary

property_name_query

See PropertyNameQuery

property_name_suggestion

See PropertyNameSuggestion

provisioning_parameter

See ProvisioningParameter

public_workforce_task_price

See PublicWorkforceTaskPrice

redshift_dataset_definition

See RedshiftDatasetDefinition

register_model_step_metadata

See RegisterModelStepMetadata

renderable_task

See RenderableTask

rendering_error

See RenderingError

repository_auth_config

See RepositoryAuthConfig

resolved_attributes

See ResolvedAttributes

resource_config

See ResourceConfig

resource_limits

See ResourceLimits

resource_spec

See ResourceSpec

retention_policy

See RetentionPolicy

retry_strategy

See RetryStrategy

s3_data_source

See S3DataSource

s3_storage_config

See S3StorageConfig

schedule_config

See ScheduleConfig

search_expression

See SearchExpression

search_record

See SearchRecord

secondary_status_transition

See SecondaryStatusTransition

service_catalog_provisioned_product_details

See ServiceCatalogProvisionedProductDetails

service_catalog_provisioning_details

See ServiceCatalogProvisioningDetails

sharing_settings

See SharingSettings

shuffle_config

See ShuffleConfig

source_algorithm

See SourceAlgorithm

source_algorithm_specification

See SourceAlgorithmSpecification

source_ip_config

See SourceIpConfig

stopping_condition

See StoppingCondition

subscribed_workteam

See SubscribedWorkteam

suggestion_query

See SuggestionQuery

tag

See Tag

target_platform

See TargetPlatform

tensor_board_app_settings

See TensorBoardAppSettings

tensor_board_output_config

See TensorBoardOutputConfig

traffic_routing_config

See TrafficRoutingConfig

training_job

See TrainingJob

training_job_definition

See TrainingJobDefinition

training_job_status_counters

See TrainingJobStatusCounters

training_job_step_metadata

See TrainingJobStepMetadata

training_job_summary

See TrainingJobSummary

training_specification

See TrainingSpecification

transform_data_source

See TransformDataSource

transform_input

See TransformInput

transform_job

See TransformJob

transform_job_definition

See TransformJobDefinition

transform_job_step_metadata

See TransformJobStepMetadata

transform_job_summary

See TransformJobSummary

transform_output

See TransformOutput

transform_resources

See TransformResources

transform_s3_data_source

See TransformS3DataSource

trial

See Trial

trial_component

See TrialComponent

trial_component_artifact

See TrialComponentArtifact

trial_component_metric_summary

See TrialComponentMetricSummary

trial_component_simple_summary

See TrialComponentSimpleSummary

trial_component_source

See TrialComponentSource

trial_component_source_detail

See TrialComponentSourceDetail

trial_component_status

See TrialComponentStatus

trial_component_summary

See TrialComponentSummary

trial_source

See TrialSource

trial_summary

See TrialSummary

tuning_job_completion_criteria

See TuningJobCompletionCriteria

ui_config

See UiConfig

ui_template

See UiTemplate

ui_template_info

See UiTemplateInfo

usd

See Usd

user_context

See UserContext

user_profile_details

See UserProfileDetails

user_settings

See UserSettings

variant_property

See VariantProperty

vpc_config

See VpcConfig

workforce

See Workforce

workteam

See Workteam

Structs

ActionSource

A structure describing the source of an action.

ActionSummary

Lists the properties of an action. An action represents an action

AgentVersion

Edge Manager agent version.

Alarm

This API is not supported.

AlgorithmSpecification

Specifies the training algorithm to use in a CreateTrainingJob

AlgorithmStatusDetails

Specifies the validation and image scan statuses of the algorithm.

AlgorithmStatusItem

Represents the overall status of an algorithm.

AlgorithmSummary

Provides summary information about an algorithm.

AlgorithmValidationProfile

Defines a training job and a batch transform job that Amazon SageMaker runs to validate your

AlgorithmValidationSpecification

Specifies configurations for one or more training jobs that Amazon SageMaker runs to test the

AnnotationConsolidationConfig

Configures how labels are consolidated across human workers and processes output data.

AppDetails

Details about an Amazon SageMaker app.

AppImageConfigDetails

The configuration for running a SageMaker image as a KernelGateway app.

AppSpecification

Configuration to run a processing job in a specified container image.

ArtifactSource

A structure describing the source of an artifact.

ArtifactSourceType

The ID and ID type of an artifact source.

ArtifactSummary

Lists a summary of the properties of an artifact. An artifact represents a URI

AssociationSummary

Lists a summary of the properties of an association. An association is an entity that

AthenaDatasetDefinition

Configuration for Athena Dataset Definition input.

AutoMlCandidate

An Autopilot job returns recommendations, or candidates. Each candidate has futher details

AutoMlCandidateStep

Information about the steps for a candidate and what step it is working on.

AutoMlChannel

A channel is a named input source that training algorithms can consume. For more

AutoMlContainerDefinition

A list of container definitions that describe the different containers that make up an

AutoMlDataSource

The data source for the Autopilot job.

AutoMlJobArtifacts

The artifacts that are generated during an AutoML job.

AutoMlJobCompletionCriteria

How long a job is allowed to run, or how many candidates a job is allowed to

AutoMlJobConfig

A collection of settings used for an AutoML job.

AutoMlJobObjective

Specifies a metric to minimize or maximize as the objective of a job.

AutoMlJobSummary

Provides a summary about an AutoML job.

AutoMlOutputDataConfig

The output data configuration.

AutoMlPartialFailureReason

The reason for a partial failure of an AutoML job.

AutoMlSecurityConfig

Security options.

AutoMls3DataSource

The Amazon S3 data source.

AutoRollbackConfig

Currently, the AutoRollbackConfig API is not supported.

Bias

Contains bias metrics for a model.

BlueGreenUpdatePolicy

Currently, the BlueGreenUpdatePolicy API is not supported.

CacheHitResult

Details on the cache hit of a pipeline execution step.

CallbackStepMetadata

Metadata about a callback step.

CandidateArtifactLocations

The location of artifacts for an AutoML candidate job.

CandidateProperties

The properties of an AutoML candidate job.

CapacitySize

Currently, the CapacitySize API is not supported.

CaptureContentTypeHeader

CaptureOption

CategoricalParameterRange

A list of categorical hyperparameters to tune.

CategoricalParameterRangeSpecification

Defines the possible values for a categorical hyperparameter.

Channel

A channel is a named input source that training algorithms can consume.

ChannelSpecification

Defines a named input source, called a channel, to be used by an algorithm.

CheckpointConfig

Contains information about the output location for managed spot training checkpoint

CodeRepositorySummary

Specifies summary information about a Git repository.

CognitoConfig

Use this parameter to configure your Amazon Cognito workforce.

CognitoMemberDefinition

Identifies a Amazon Cognito user group. A user group can be used in on or more work

CollectionConfiguration

Configuration information for the Debugger output tensor collections.

CompilationJobSummary

A summary of a model compilation job.

ConditionStepMetadata

Metadata for a Condition step.

ContainerDefinition

Describes the container, as part of model definition.

ContextSource

A structure describing the source of a context.

ContextSummary

Lists a summary of the properties of a context. A context provides a logical grouping

ContinuousParameterRange

A list of continuous hyperparameters to tune.

ContinuousParameterRangeSpecification

Defines the possible values for a continuous hyperparameter.

CustomImage

A custom SageMaker image. For more information, see

DataCaptureConfig

DataCaptureConfigSummary

DataCatalogConfig

The meta data of the Glue table which serves as data catalog for the

DataProcessing

The data structure used to specify the data to be used for inference in a batch

DataQualityAppSpecification

Information about the container that a data quality monitoring job runs.

DataQualityBaselineConfig

Configuration for monitoring constraints and monitoring statistics. These baseline

DataQualityJobInput

The input for the data quality monitoring job. Currently endpoints are supported for

DataSource

Describes the location of the channel data.

DatasetDefinition

Configuration for Dataset Definition inputs. The Dataset Definition input must specify

DebugHookConfig

Configuration information for the Debugger hook parameters, metric and tensor collections, and

DebugRuleConfiguration

Configuration information for SageMaker Debugger rules for debugging. To learn more about

DebugRuleEvaluationStatus

Information about the status of the rule evaluation.

DeployedImage

Gets the Amazon EC2 Container Registry path of the docker image of the model that is hosted in this ProductionVariant.

DeploymentConfig

Currently, the DeploymentConfig API is not supported.

DesiredWeightAndCapacity

Specifies weight and capacity values for a production variant.

Device

Information of a particular device.

DeviceFleetSummary

Summary of the device fleet.

DeviceStats

Status of devices.

DeviceSummary

Summary of the device.

DomainDetails

The domain's details.

EdgeModel

The model on the edge device.

EdgeModelStat

Status of edge devices with this model.

EdgeModelSummary

Summary of model on edge device.

EdgeOutputConfig

The output configuration.

EdgePackagingJobSummary

Summary of edge packaging job.

EdgePresetDeploymentOutput

The output of a SageMaker Edge Manager deployable resource.

Endpoint

A hosted endpoint for real-time inference.

EndpointConfigSummary

Provides summary information for an endpoint configuration.

EndpointInput

Input object for the endpoint

EndpointSummary

Provides summary information for an endpoint.

Experiment

The properties of an experiment as returned by the Search API.

ExperimentConfig

Associates a SageMaker job as a trial component with an experiment and trial. Specified when

ExperimentSource

The source of the experiment.

ExperimentSummary

A summary of the properties of an experiment. To get the complete set of properties, call

Explainability

Contains explainability metrics for a model.

FeatureDefinition

A list of features. You must include FeatureName and

FeatureGroup

Amazon SageMaker Feature Store stores features in a collection called Feature Group.

FeatureGroupSummary

The name, Arn, CreationTime, FeatureGroup values,

FileSystemConfig

The Amazon Elastic File System (EFS) storage configuration for a SageMaker image.

FileSystemDataSource

Specifies a file system data source for a channel.

Filter

A conditional statement for a search expression that includes a resource property, a

FinalAutoMlJobObjectiveMetric

The best candidate result from an AutoML training job.

FinalHyperParameterTuningJobObjectiveMetric

Shows the final value for the

FlowDefinitionOutputConfig

Contains information about where human output will be stored.

FlowDefinitionSummary

Contains summary information about the flow definition.

GitConfig

Specifies configuration details for a Git repository in your Amazon Web Services account.

GitConfigForUpdate

Specifies configuration details for a Git repository when the repository is

HumanLoopActivationConditionsConfig

Defines under what conditions SageMaker creates a human loop. Used within . See for the required

HumanLoopActivationConfig

Provides information about how and under what conditions SageMaker creates a human loop. If HumanLoopActivationConfig is not given, then all requests go to humans.

HumanLoopConfig

Describes the work to be performed by human workers.

HumanLoopRequestSource

Container for configuring the source of human task requests.

HumanTaskConfig

Information required for human workers to complete a labeling task.

HumanTaskUiSummary

Container for human task user interface information.

HyperParameterAlgorithmSpecification

Specifies

HyperParameterSpecification

Defines a hyperparameter to be used by an algorithm.

HyperParameterTrainingJobDefinition

Defines

HyperParameterTrainingJobSummary

Specifies

HyperParameterTuningJobConfig

Configures a hyperparameter tuning job.

HyperParameterTuningJobObjective

Defines the objective metric for a hyperparameter tuning job.

HyperParameterTuningJobSummary

Provides summary information about a hyperparameter tuning job.

HyperParameterTuningJobWarmStartConfig

Specifies the configuration for a hyperparameter tuning job that uses one or more

Image

A SageMaker image. A SageMaker image represents a set of container images that are derived from

ImageConfig

Specifies whether the model container is in Amazon ECR or a private Docker registry

ImageVersion

A version of a SageMaker Image. A version represents an existing container

InferenceExecutionConfig

Specifies details about how containers in a multi-container endpoint are run.

InferenceSpecification

Defines how to perform inference generation after a training job is run.

InputConfig

Contains information about the location of input model artifacts, the name and

IntegerParameterRange

For a hyperparameter of the integer type, specifies the range

IntegerParameterRangeSpecification

Defines the possible values for an integer hyperparameter.

JupyterServerAppSettings

The JupyterServer app settings.

KernelGatewayAppSettings

The KernelGateway app settings.

KernelGatewayImageConfig

The configuration for the file system and kernels in a SageMaker image running as a

KernelSpec

The specification of a Jupyter kernel.

LabelCounters

Provides a breakdown of the number of objects labeled.

LabelCountersForWorkteam

Provides counts for human-labeled tasks in the labeling job.

LabelingJobAlgorithmsConfig

Provides configuration information for auto-labeling of your data objects. A

LabelingJobDataAttributes

Attributes of the data specified by the customer. Use these to describe the data to be

LabelingJobDataSource

Provides information about the location of input data.

LabelingJobForWorkteamSummary

Provides summary information for a work team.

LabelingJobInputConfig

Input configuration information for a labeling job.

LabelingJobOutput

Specifies the location of the output produced by the labeling job.

LabelingJobOutputConfig

Output configuration information for a labeling job.

LabelingJobResourceConfig

Configure encryption on the storage volume attached to the ML compute instance used to

LabelingJobS3DataSource

The Amazon S3 location of the input data objects.

LabelingJobSnsDataSource

An Amazon SNS data source used for streaming labeling jobs.

LabelingJobStoppingConditions

A set of conditions for stopping a labeling job. If any of the conditions are met, the

LabelingJobSummary

Provides summary information about a labeling job.

MemberDefinition

Defines an Amazon Cognito or your own OIDC IdP user group that is part of a work team.

MetadataProperties

Metadata properties of the tracking entity, trial, or trial component.

MetricData

The name, value, and date and time of a metric that was emitted to Amazon CloudWatch.

MetricDefinition

Specifies a metric that the training algorithm

MetricsSource

ModelArtifacts

Provides information about the location that is configured for storing model

ModelBiasAppSpecification

Docker container image configuration object for the model bias job.

ModelBiasBaselineConfig

The configuration for a baseline model bias job.

ModelBiasJobInput

Inputs for the model bias job.

ModelClientConfig

Configures the timeout and maximum number of retries for processing a transform job

ModelDataQuality

Data quality constraints and statistics for a model.

ModelDeployConfig

Specifies how to generate the endpoint name for an automatic one-click Autopilot model

ModelDeployResult

Provides information about the endpoint of the model deployment.

ModelDigests

Provides information to verify the integrity of stored model artifacts.

ModelExplainabilityAppSpecification

Docker container image configuration object for the model explainability job.

ModelExplainabilityBaselineConfig

The configuration for a baseline model explainability job.

ModelExplainabilityJobInput

Inputs for the model explainability job.

ModelMetrics

Contains metrics captured from a model.

ModelPackage

A versioned model that can be deployed for SageMaker inference.

ModelPackageContainerDefinition

Describes the Docker container for the model package.

ModelPackageGroup

A group of versioned models in the model registry.

ModelPackageGroupSummary

Summary information about a model group.

ModelPackageStatusDetails

Specifies the validation and image scan statuses of the model package.

ModelPackageStatusItem

Represents the overall status of a model package.

ModelPackageSummary

Provides summary information about a model package.

ModelPackageValidationProfile

Contains data, such as the inputs and targeted instance types that are used in the

ModelPackageValidationSpecification

Specifies batch transform jobs that Amazon SageMaker runs to validate your model package.

ModelQuality

Model quality statistics and constraints.

ModelQualityAppSpecification

Container image configuration object for the monitoring job.

ModelQualityBaselineConfig

Configuration for monitoring constraints and monitoring statistics. These baseline

ModelQualityJobInput

The input for the model quality monitoring job. Currently endponts are supported for

ModelStepMetadata

Metadata for Model steps.

ModelSummary

Provides summary information about a model.

MonitoringAppSpecification

Container image configuration object for the monitoring job.

MonitoringBaselineConfig

Configuration for monitoring constraints and monitoring statistics. These baseline

MonitoringClusterConfig

Configuration for the cluster used to run model monitoring jobs.

MonitoringConstraintsResource

The constraints resource for a monitoring job.

MonitoringExecutionSummary

Summary of information about the last monitoring job to run.

MonitoringGroundTruthS3Input

The ground truth labels for the dataset used for the monitoring job.

MonitoringInput

The inputs for a monitoring job.

MonitoringJobDefinition

Defines the monitoring job.

MonitoringJobDefinitionSummary

Summary information about a monitoring job.

MonitoringNetworkConfig

The networking configuration for the monitoring job.

MonitoringOutput

The output object for a monitoring job.

MonitoringOutputConfig

The output configuration for monitoring jobs.

MonitoringResources

Identifies the resources to deploy for a monitoring job.

MonitoringS3Output

Information about where and how you want to store the results of a monitoring

MonitoringSchedule

A schedule for a model monitoring job. For information about model monitor, see

MonitoringScheduleConfig

Configures the monitoring schedule and defines the monitoring job.

MonitoringScheduleSummary

Summarizes the monitoring schedule.

MonitoringStatisticsResource

The statistics resource for a monitoring job.

MonitoringStoppingCondition

A time limit for how long the monitoring job is allowed to run before stopping.

MultiModelConfig

Specifies additional configuration for hosting multi-model endpoints.

NeoVpcConfig

The VpcConfig configuration object that specifies the VPC that you

NestedFilters

A list of nested Filter objects. A resource must satisfy the conditions

NetworkConfig

Networking options for a job, such as network traffic encryption between containers,

NotebookInstanceLifecycleConfigSummary

Provides a summary of a notebook instance lifecycle configuration.

NotebookInstanceLifecycleHook

Contains the notebook instance lifecycle configuration script.

NotebookInstanceSummary

Provides summary information for an Amazon SageMaker notebook instance.

NotificationConfiguration

Configures SNS notifications of available or expiring work items for work

ObjectiveStatusCounters

Specifies the number of training jobs that this hyperparameter tuning job launched,

OfflineStoreConfig

The configuration of an OfflineStore.

OfflineStoreStatus

The status of OfflineStore.

OidcConfig

Use this parameter to configure your OIDC Identity Provider (IdP).

OidcConfigForResponse

Your OIDC IdP workforce configuration.

OidcMemberDefinition

A list of user groups that exist in your OIDC Identity Provider (IdP).

OnlineStoreConfig

Use this to specify the Amazon Web Services Key Management Service (KMS) Key ID, or

OnlineStoreSecurityConfig

The security configuration for OnlineStore.

OutputConfig

Contains information about the output location for the compiled model and the target

OutputDataConfig

Provides information about how to store model training results (model

OutputParameter

An output parameter of a pipeline step.

Parameter

Assigns a value to a named Pipeline parameter.

ParameterRange

Defines the possible values for categorical, continuous, and integer hyperparameters

ParameterRanges

Specifies ranges of integer, continuous, and categorical hyperparameters that a

Parent

The trial that a trial component is associated with and the experiment the trial is part

ParentHyperParameterTuningJob

A previously completed or stopped hyperparameter tuning job to be used as a starting

Pipeline

A SageMaker Model Building Pipeline instance.

PipelineExecution

An execution of a pipeline.

PipelineExecutionStep

An execution of a step in a pipeline.

PipelineExecutionStepMetadata

Metadata for a step execution.

PipelineExecutionSummary

A pipeline execution summary.

PipelineExperimentConfig

Specifies the names of the experiment and trial created by a pipeline.

PipelineSummary

A summary of a pipeline.

ProcessingClusterConfig

Configuration for the cluster used to run a processing job.

ProcessingFeatureStoreOutput

Configuration for processing job outputs in Amazon SageMaker Feature Store.

ProcessingInput

The inputs for a processing job. The processing input must specify exactly one of either

ProcessingJob

An Amazon SageMaker processing job that is used to analyze data and evaluate models. For more information,

ProcessingJobStepMetadata

Metadata for a processing job step.

ProcessingJobSummary

Summary of information about a processing job.

ProcessingOutput

Describes the results of a processing job. The processing output must specify exactly one of

ProcessingOutputConfig

Configuration for uploading output from the processing container.

ProcessingResources

Identifies the resources, ML compute instances, and ML storage volumes to deploy for a

ProcessingS3Input

Configuration for downloading input data from Amazon S3 into the processing container.

ProcessingS3Output

Configuration for uploading output data to Amazon S3 from the processing container.

ProcessingStoppingCondition

Configures conditions under which the processing job should be stopped, such as how long

ProductionVariant

Identifies a model that you want to host and the resources chosen to deploy for

ProductionVariantCoreDumpConfig

Specifies configuration for a core dump from the model container when the process

ProductionVariantSummary

Describes weight and capacities for a production variant associated with an

ProfilerConfig

Configuration information for Debugger system monitoring, framework profiling, and

ProfilerConfigForUpdate

Configuration information for updating the Debugger profile parameters, system and framework metrics configurations, and

ProfilerRuleConfiguration

Configuration information for profiling rules.

ProfilerRuleEvaluationStatus

Information about the status of the rule evaluation.

ProjectSummary

Information about a project.

PropertyNameQuery

Part of the SuggestionQuery type. Specifies a hint for retrieving property

PropertyNameSuggestion

A property name returned from a GetSearchSuggestions call that specifies

ProvisioningParameter

A key value pair used when you provision a project as a service catalog product. For

PublicWorkforceTaskPrice

Defines the amount of money paid to an Amazon Mechanical Turk worker for each task performed.

RedshiftDatasetDefinition

Configuration for Redshift Dataset Definition input.

RegisterModelStepMetadata

Metadata for a register model job step.

RenderableTask

Contains input values for a task.

RenderingError

A description of an error that occurred while rendering the template.

RepositoryAuthConfig

Specifies an authentication configuration for the private docker registry where your

ResolvedAttributes

The resolved attributes.

ResourceConfig

Describes the resources, including ML compute instances and ML storage volumes, to

ResourceLimits

Specifies the maximum number of

ResourceSpec

Specifies the ARN's of a SageMaker image and SageMaker image version, and the instance type that

RetentionPolicy

The retention policy for data stored on an Amazon Elastic File System (EFS) volume.

RetryStrategy

The retry strategy to use when a training job fails due to an

S3DataSource

Describes the S3 data source.

S3StorageConfig

The Amazon Simple Storage (Amazon S3) location and and security configuration for OfflineStore.

ScheduleConfig

Configuration details about the monitoring schedule.

SearchExpression

A multi-expression that searches for the specified resource or resources in a search. All resource

SearchRecord

A single resource returned as part of the Search API response.

SecondaryStatusTransition

An array element of DescribeTrainingJobResponse$SecondaryStatusTransitions. It provides

ServiceCatalogProvisionedProductDetails

Details of a provisioned service catalog product. For information about service catalog,

ServiceCatalogProvisioningDetails

Details that you specify to provision a service catalog product. For information about

SharingSettings

Specifies options for sharing SageMaker Studio notebooks. These settings are

ShuffleConfig

A configuration for a shuffle option for input data in a channel. If you use

SourceAlgorithm

Specifies an algorithm that was used to create the model package. The algorithm must

SourceAlgorithmSpecification

A list of algorithms that were used to create a model package.

SourceIpConfig

A list of IP address ranges (CIDRs). Used to create an allow

StoppingCondition

Specifies a limit to how long a model training job, model compilation job, or

SubscribedWorkteam

Describes a work team of a vendor that does the a labelling job.

SuggestionQuery

Specified in the GetSearchSuggestions request.

Tag

A tag object that consists of a key and an optional value, used to manage metadata

TargetPlatform

Contains information about a target platform that you want your model to run on, such

TensorBoardAppSettings

The TensorBoard app settings.

TensorBoardOutputConfig

Configuration of storage locations for the Debugger TensorBoard output data.

TrafficRoutingConfig

Currently, the TrafficRoutingConfig API is not supported.

TrainingJob

Contains information about a training job.

TrainingJobDefinition

Defines the input needed to run a training job using the algorithm.

TrainingJobStatusCounters

The numbers of training jobs launched by a hyperparameter tuning job, categorized by

TrainingJobStepMetadata

Metadata for a training job step.

TrainingJobSummary

Provides summary information about a training job.

TrainingSpecification

Defines how the algorithm is used for a training job.

TransformDataSource

Describes the location of the channel data.

TransformInput

Describes the input source of a transform job and the way the transform job consumes

TransformJob

A batch transform job. For information about SageMaker batch transform, see Use Batch

TransformJobDefinition

Defines the input needed to run a transform job using the inference specification

TransformJobStepMetadata

Metadata for a transform job step.

TransformJobSummary

Provides a

TransformOutput

Describes the results of a transform job.

TransformResources

Describes the resources, including ML instance types and ML instance count, to use for

TransformS3DataSource

Describes the S3 data source.

Trial

The properties of a trial as returned by the Search API.

TrialComponent

The properties of a trial component as returned by the Search

TrialComponentArtifact

Represents an input or output artifact of a trial component. You specify

TrialComponentMetricSummary

A summary of the metrics of a trial component.

TrialComponentSimpleSummary

A short summary of a trial component.

TrialComponentSource

The Amazon Resource Name (ARN) and job type of the source of a trial component.

TrialComponentSourceDetail

Detailed information about the source of a trial component. Either

TrialComponentStatus

The status of the trial component.

TrialComponentSummary

A summary of the properties of a trial component. To get all the properties, call the

TrialSource

The source of the trial.

TrialSummary

A summary of the properties of a trial. To get the complete set of properties, call the

TuningJobCompletionCriteria

The job completion criteria.

UiConfig

Provided configuration information for the worker UI for a labeling job.

UiTemplate

The Liquid template for the worker user interface.

UiTemplateInfo

Container for user interface template information.

Usd

Represents an amount of money in United States dollars.

UserContext

Information about the user who created or modified an experiment, trial, or trial

UserProfileDetails

The user profile details.

UserSettings

A collection of settings that apply to users of Amazon SageMaker Studio. These settings are

VariantProperty

Specifies a production variant property type for an Endpoint.

VpcConfig

Specifies a VPC that your training jobs and hosted models have access to. Control

Workforce

A single private workforce, which is automatically created when you create your first

Workteam

Provides details about a labeling work team.

Enums

ActionStatus

NOTE: ActionStatus::Unknown has been renamed to ::UnknownValue.

AlgorithmSortBy

AlgorithmStatus

AppImageConfigSortKey

AppInstanceType

AppNetworkAccessType

AppSortKey

AppStatus

AppType

ArtifactSourceIdType

AssemblyType

AssociationEdgeType

AthenaResultCompressionType

The compression used for Athena query results.

AthenaResultFormat

The data storage format for Athena query results.

AuthMode

AutoMlJobObjectiveType

AutoMlJobSecondaryStatus

AutoMlJobStatus

AutoMlMetricEnum

AutoMlSortBy

AutoMlSortOrder

AutoMls3DataType

AwsManagedHumanLoopRequestSource

BatchStrategy

BooleanOperator

CandidateSortBy

CandidateStatus

CandidateStepType

CapacitySizeType

CaptureMode

CaptureStatus

CodeRepositorySortBy

CodeRepositorySortOrder

CompilationJobStatus

CompressionType

ConditionOutcome

ContainerMode

ContentClassifier

DataDistributionType

DetailedAlgorithmStatus

DetailedModelPackageStatus

DirectInternetAccess

DomainStatus

EdgePackagingJobStatus

EdgePresetDeploymentStatus

EdgePresetDeploymentType

EndpointConfigSortKey

EndpointSortKey

EndpointStatus

ExecutionStatus

FeatureGroupSortBy

FeatureGroupSortOrder

FeatureGroupStatus

FeatureType

FileSystemAccessMode

FileSystemType

FlowDefinitionStatus

Framework

HumanTaskUiStatus

HyperParameterScalingType

HyperParameterTuningJobObjectiveType

HyperParameterTuningJobSortByOptions

HyperParameterTuningJobStatus

HyperParameterTuningJobStrategyType

The strategy hyperparameter tuning uses to

HyperParameterTuningJobWarmStartType

ImageSortBy

ImageSortOrder

ImageStatus

ImageVersionSortBy

ImageVersionSortOrder

ImageVersionStatus

InferenceExecutionMode

InputMode

InstanceType

JoinSource

LabelingJobStatus

ListCompilationJobsSortBy

ListDeviceFleetsSortBy

ListEdgePackagingJobsSortBy

ListLabelingJobsForWorkteamSortByOptions

ListWorkforcesSortByOptions

ListWorkteamsSortByOptions

ModelApprovalStatus

ModelCacheSetting

ModelPackageGroupSortBy

ModelPackageGroupStatus

ModelPackageSortBy

ModelPackageStatus

ModelPackageType

ModelSortKey

MonitoringExecutionSortKey

MonitoringJobDefinitionSortKey

MonitoringProblemType

MonitoringScheduleSortKey

MonitoringType

NotebookInstanceAcceleratorType

NotebookInstanceLifecycleConfigSortKey

NotebookInstanceLifecycleConfigSortOrder

NotebookInstanceSortKey

NotebookInstanceSortOrder

NotebookInstanceStatus

NotebookOutputOption

ObjectiveStatus

OfflineStoreStatusValue

Operator

OrderKey

ParameterType

PipelineExecutionStatus

PipelineStatus

ProblemType

ProcessingInstanceType

ProcessingJobStatus

ProcessingS3CompressionType

ProcessingS3DataDistributionType

ProcessingS3DataType

ProcessingS3InputMode

ProcessingS3UploadMode

ProductionVariantAcceleratorType

ProductionVariantInstanceType

ProfilingStatus

ProjectSortBy

ProjectSortOrder

ProjectStatus

RecordWrapper

RedshiftResultCompressionType

The compression used for Redshift query results.

RedshiftResultFormat

The data storage format for Redshift query results.

RepositoryAccessMode

ResourceType

RetentionType

RootAccess

RuleEvaluationStatus

S3DataDistribution

S3DataType

SagemakerServicecatalogStatus

ScheduleStatus

SearchSortOrder

SecondaryStatus

SortActionsBy

SortArtifactsBy

SortAssociationsBy

SortBy

SortContextsBy

SortExperimentsBy

SortOrder

SortPipelineExecutionsBy

SortPipelinesBy

SortTrialComponentsBy

SortTrialsBy

SplitType

StepStatus

TargetDevice

TargetPlatformAccelerator

TargetPlatformArch

TargetPlatformOs

TrafficRoutingConfigType

TrainingInputMode

TrainingInstanceType

TrainingJobEarlyStoppingType

TrainingJobSortByOptions

TrainingJobStatus

TransformInstanceType

TransformJobStatus

TrialComponentParameterValue

TrialComponentPrimaryStatus

UserProfileSortKey

UserProfileStatus

VariantPropertyType