Struct aws_sdk_sagemaker::model::debug_rule_configuration::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for DebugRuleConfiguration

Implementations

impl Builder

pub fn rule_configuration_name(self, input: impl Into<String>) -> Self

The name of the rule configuration. It must be unique relative to other rule configuration names.

pub fn set_rule_configuration_name(self, input: Option<String>) -> Self

pub fn local_path(self, input: impl Into<String>) -> Self

Path to local storage location for output of rules. Defaults to /opt/ml/processing/output/rule/.

pub fn set_local_path(self, input: Option<String>) -> Self

pub fn s3_output_path(self, input: impl Into<String>) -> Self

Path to Amazon S3 storage location for rules.

pub fn set_s3_output_path(self, input: Option<String>) -> Self

pub fn rule_evaluator_image(self, input: impl Into<String>) -> Self

The Amazon Elastic Container (ECR) Image for the managed rule evaluation.

pub fn set_rule_evaluator_image(self, input: Option<String>) -> Self

pub fn instance_type(self, input: ProcessingInstanceType) -> Self

The instance type to deploy a Debugger custom rule for debugging a training job.

pub fn set_instance_type(self, input: Option<ProcessingInstanceType>) -> Self

pub fn volume_size_in_gb(self, input: i32) -> Self

The size, in GB, of the ML storage volume attached to the processing instance.

pub fn set_volume_size_in_gb(self, input: Option<i32>) -> Self

pub fn rule_parameters(self, k: impl Into<String>, v: impl Into<String>) -> Self

pub fn set_rule_parameters(self, input: Option<HashMap<String, String>>) -> Self

pub fn build(self) -> DebugRuleConfiguration

Consumes the builder and constructs a DebugRuleConfiguration