Struct aws_sdk_sagemaker::model::hyper_parameter_tuning_job_config::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for HyperParameterTuningJobConfig

Implementations

impl Builder

pub fn strategy(self, input: HyperParameterTuningJobStrategyType) -> Self

Specifies how hyperparameter tuning chooses the combinations of hyperparameter values to use for the training job it launches. To use the Bayesian search strategy, set this to Bayesian. To randomly search, set it to Random. For information about search strategies, see How Hyperparameter Tuning Works.

pub fn set_strategy(
self,
input: Option<HyperParameterTuningJobStrategyType>
) -> Self

pub fn hyper_parameter_tuning_job_objective(
self,
input: HyperParameterTuningJobObjective
) -> Self

The HyperParameterTuningJobObjective object that specifies the objective metric for this tuning job.

pub fn set_hyper_parameter_tuning_job_objective(
self,
input: Option<HyperParameterTuningJobObjective>
) -> Self

pub fn resource_limits(self, input: ResourceLimits) -> Self

The ResourceLimits object that specifies the maximum number of training jobs and parallel training jobs for this tuning job.

pub fn set_resource_limits(self, input: Option<ResourceLimits>) -> Self

pub fn parameter_ranges(self, input: ParameterRanges) -> Self

The ParameterRanges object that specifies the ranges of hyperparameters that this tuning job searches.

pub fn set_parameter_ranges(self, input: Option<ParameterRanges>) -> Self

pub fn training_job_early_stopping_type(
self,
input: TrainingJobEarlyStoppingType
) -> Self

Specifies whether to use early stopping for training jobs launched by the hyperparameter tuning job. This can be one of the following values (the default value is OFF):

OFF

Training jobs launched by the hyperparameter tuning job do not use early stopping.

AUTO

Amazon SageMaker stops training jobs launched by the hyperparameter tuning job when they are unlikely to perform better than previously completed training jobs. For more information, see Stop Training Jobs Early.

pub fn set_training_job_early_stopping_type(
self,
input: Option<TrainingJobEarlyStoppingType>
) -> Self

pub fn tuning_job_completion_criteria(
self,
input: TuningJobCompletionCriteria
) -> Self

The tuning job's completion criteria.

pub fn set_tuning_job_completion_criteria(
self,
input: Option<TuningJobCompletionCriteria>
) -> Self

pub fn build(self) -> HyperParameterTuningJobConfig

Consumes the builder and constructs a HyperParameterTuningJobConfig