Struct aws_sdk_sagemaker::model::monitoring_job_definition::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for MonitoringJobDefinition

Implementations

impl Builder

pub fn baseline_config(self, input: MonitoringBaselineConfig) -> Self

Baseline configuration used to validate that the data conforms to the specified constraints and statistics

pub fn set_baseline_config(
self,
input: Option<MonitoringBaselineConfig>
) -> Self

pub fn monitoring_inputs(self, input: impl Into<MonitoringInput>) -> Self

pub fn set_monitoring_inputs(self, input: Option<Vec<MonitoringInput>>) -> Self

pub fn monitoring_output_config(self, input: MonitoringOutputConfig) -> Self

The array of outputs from the monitoring job to be uploaded to Amazon Simple Storage Service (Amazon S3).

pub fn set_monitoring_output_config(
self,
input: Option<MonitoringOutputConfig>
) -> Self

pub fn monitoring_resources(self, input: MonitoringResources) -> Self

Identifies the resources, ML compute instances, and ML storage volumes to deploy for a monitoring job. In distributed processing, you specify more than one instance.

pub fn set_monitoring_resources(
self,
input: Option<MonitoringResources>
) -> Self

pub fn monitoring_app_specification(
self,
input: MonitoringAppSpecification
) -> Self

Configures the monitoring job to run a specified Docker container image.

pub fn set_monitoring_app_specification(
self,
input: Option<MonitoringAppSpecification>
) -> Self

pub fn stopping_condition(self, input: MonitoringStoppingCondition) -> Self

Specifies a time limit for how long the monitoring job is allowed to run.

pub fn set_stopping_condition(
self,
input: Option<MonitoringStoppingCondition>
) -> Self

pub fn environment(self, k: impl Into<String>, v: impl Into<String>) -> Self

pub fn set_environment(self, input: Option<HashMap<String, String>>) -> Self

pub fn network_config(self, input: NetworkConfig) -> Self

Specifies networking options for an monitoring job.

pub fn set_network_config(self, input: Option<NetworkConfig>) -> Self

pub fn role_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of an IAM role that Amazon SageMaker can assume to perform tasks on your behalf.

pub fn set_role_arn(self, input: Option<String>) -> Self

pub fn build(self) -> MonitoringJobDefinition

Consumes the builder and constructs a MonitoringJobDefinition