Struct aws_sdk_sagemaker::model::TransformJobDefinition

#[non_exhaustive]
pub struct TransformJobDefinition {
    pub max_concurrent_transforms: Option<i32>,
    pub max_payload_in_mb: Option<i32>,
    pub batch_strategy: Option<BatchStrategy>,
    pub environment: Option<HashMap<String, String>>,
    pub transform_input: Option<TransformInput>,
    pub transform_output: Option<TransformOutput>,
    pub transform_resources: Option<TransformResources>,
}

Defines the input needed to run a transform job using the inference specification specified in the algorithm.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

max_concurrent_transforms: Option<i32>

The maximum number of parallel requests that can be sent to each instance in a transform job. The default value is 1.

max_payload_in_mb: Option<i32>

The maximum payload size allowed, in MB. A payload is the data portion of a record (without metadata).

batch_strategy: Option<BatchStrategy>

A string that determines the number of records included in a single mini-batch.

SingleRecord means only one record is used per mini-batch. MultiRecord means a mini-batch is set to contain as many records that can fit within the MaxPayloadInMB limit.

environment: Option<HashMap<String, String>>

The environment variables to set in the Docker container. We support up to 16 key and values entries in the map.

transform_input: Option<TransformInput>

A description of the input source and the way the transform job consumes it.

transform_output: Option<TransformOutput>

Identifies the Amazon S3 location where you want Amazon SageMaker to save the results from the transform job.

transform_resources: Option<TransformResources>

Identifies the ML compute instances for the transform job.

Implementations

impl TransformJobDefinition

pub fn builder() -> Builder

Creates a new builder-style object to manufacture TransformJobDefinition