Struct aws_sdk_sagemaker::model::trial_component::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for TrialComponent

Implementations

impl Builder

pub fn trial_component_name(self, input: impl Into<String>) -> Self

The name of the trial component.

pub fn set_trial_component_name(self, input: Option<String>) -> Self

pub fn display_name(self, input: impl Into<String>) -> Self

The name of the component as displayed. If DisplayName isn't specified, TrialComponentName is displayed.

pub fn set_display_name(self, input: Option<String>) -> Self

pub fn trial_component_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the trial component.

pub fn set_trial_component_arn(self, input: Option<String>) -> Self

pub fn source(self, input: TrialComponentSource) -> Self

The Amazon Resource Name (ARN) and job type of the source of the component.

pub fn set_source(self, input: Option<TrialComponentSource>) -> Self

pub fn status(self, input: TrialComponentStatus) -> Self

The status of the trial component.

pub fn set_status(self, input: Option<TrialComponentStatus>) -> Self

pub fn start_time(self, input: Instant) -> Self

When the component started.

pub fn set_start_time(self, input: Option<Instant>) -> Self

pub fn end_time(self, input: Instant) -> Self

When the component ended.

pub fn set_end_time(self, input: Option<Instant>) -> Self

pub fn creation_time(self, input: Instant) -> Self

When the component was created.

pub fn set_creation_time(self, input: Option<Instant>) -> Self

pub fn created_by(self, input: UserContext) -> Self

Information about the user who created or modified an experiment, trial, or trial component.

pub fn set_created_by(self, input: Option<UserContext>) -> Self

pub fn last_modified_time(self, input: Instant) -> Self

When the component was last modified.

pub fn set_last_modified_time(self, input: Option<Instant>) -> Self

pub fn last_modified_by(self, input: UserContext) -> Self

Information about the user who created or modified an experiment, trial, or trial component.

pub fn set_last_modified_by(self, input: Option<UserContext>) -> Self

pub fn parameters(
self,
k: impl Into<String>,
v: impl Into<TrialComponentParameterValue>
) -> Self

pub fn set_parameters(
self,
input: Option<HashMap<String, TrialComponentParameterValue>>
) -> Self

pub fn input_artifacts(
self,
k: impl Into<String>,
v: impl Into<TrialComponentArtifact>
) -> Self

pub fn set_input_artifacts(
self,
input: Option<HashMap<String, TrialComponentArtifact>>
) -> Self

pub fn output_artifacts(
self,
k: impl Into<String>,
v: impl Into<TrialComponentArtifact>
) -> Self

pub fn set_output_artifacts(
self,
input: Option<HashMap<String, TrialComponentArtifact>>
) -> Self

pub fn metrics(self, input: impl Into<TrialComponentMetricSummary>) -> Self

pub fn set_metrics(
self,
input: Option<Vec<TrialComponentMetricSummary>>
) -> Self

pub fn metadata_properties(self, input: MetadataProperties) -> Self

Metadata properties of the tracking entity, trial, or trial component.

pub fn set_metadata_properties(self, input: Option<MetadataProperties>) -> Self

pub fn source_detail(self, input: TrialComponentSourceDetail) -> Self

Details of the source of the component.

pub fn set_source_detail(
self,
input: Option<TrialComponentSourceDetail>
) -> Self

pub fn tags(self, input: impl Into<Tag>) -> Self

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn parents(self, input: impl Into<Parent>) -> Self

pub fn set_parents(self, input: Option<Vec<Parent>>) -> Self

pub fn build(self) -> TrialComponent

Consumes the builder and constructs a TrialComponent