Module aws_sdk_sagemaker::operation

Structs

AddAssociation

Creates an association between the source and the destination. A

AddTags

Adds or overwrites one or more tags for the specified Amazon SageMaker resource. You can add

AssociateTrialComponent

Associates a trial component with a trial. A trial component can be associated with

CreateAction

Creates an action. An action is a lineage tracking entity that

CreateAlgorithm

Create a machine learning algorithm that you can use in Amazon SageMaker and list in the Amazon Web Services

CreateApp

Creates a running app for the specified UserProfile. Supported apps are

CreateAppImageConfig

Creates a configuration for running a SageMaker image as a KernelGateway app. The

CreateArtifact

Creates an artifact. An artifact is a lineage tracking entity that

CreateAutoMLJob

Creates an Autopilot job.

CreateCodeRepository

Creates a Git repository as a resource in your Amazon SageMaker account. You can associate the

CreateCompilationJob

Starts a model compilation job. After the model has been compiled, Amazon SageMaker saves the

CreateContext

Creates a context. A context is a lineage tracking entity that

CreateDataQualityJobDefinition

Creates a definition for a job that monitors data quality and drift. For information

CreateDeviceFleet

Creates a device fleet.

CreateDomain

Creates a Domain used by Amazon SageMaker Studio. A domain consists of an associated

CreateEdgePackagingJob

Starts a SageMaker Edge Manager model packaging job. Edge Manager will use the model artifacts from the Amazon Simple Storage Service bucket that you specify. After the model has been packaged, Amazon SageMaker saves the resulting artifacts to an S3 bucket that you specify.

CreateEndpoint

Creates an endpoint using the endpoint configuration specified in the request. Amazon SageMaker

CreateEndpointConfig

Creates an endpoint configuration that Amazon SageMaker hosting services uses to deploy models. In

CreateExperiment

Creates an SageMaker experiment. An experiment is a collection of

CreateFeatureGroup

Create a new FeatureGroup. A FeatureGroup is a group of

CreateFlowDefinition

Creates a flow definition.

CreateHumanTaskUi

Defines the settings you will use for the human review workflow user interface. Reviewers will see a three-panel interface with an instruction area, the item to review, and an input area.

CreateHyperParameterTuningJob

Starts a hyperparameter tuning job. A hyperparameter tuning job finds the best version

CreateImage

Creates a custom SageMaker image. A SageMaker image is a set of image versions. Each image

CreateImageVersion

Creates a version of the SageMaker image specified by ImageName. The version

CreateLabelingJob

Creates a job that uses workers to label the data objects in your input dataset. You

CreateModel

Creates a model in Amazon SageMaker. In the request, you name the model and describe a primary

CreateModelBiasJobDefinition

Creates the definition for a model bias job.

CreateModelExplainabilityJobDefinition

Creates the definition for a model explainability job.

CreateModelPackage

Creates a model package that you can use to create Amazon SageMaker models or list on Amazon Web Services

CreateModelPackageGroup

Creates a model group. A model group contains a group of model versions.

CreateModelQualityJobDefinition

Creates a definition for a job that monitors model quality and drift. For information

CreateMonitoringSchedule

Creates a schedule that regularly starts Amazon SageMaker Processing Jobs to monitor the data

CreateNotebookInstance

Creates an Amazon SageMaker notebook instance. A notebook instance is a machine learning (ML)

CreateNotebookInstanceLifecycleConfig

Creates a lifecycle configuration that you can associate with a notebook instance. A

CreatePipeline

Creates a pipeline using a JSON pipeline definition.

CreatePresignedDomainUrl

Creates a URL for a specified UserProfile in a Domain. When accessed in a web browser,

CreatePresignedNotebookInstanceUrl

Returns a URL that you can use to connect to the Jupyter server from a notebook

CreateProcessingJob

Creates a processing job.

CreateProject

Creates a machine learning (ML) project that can contain one or more templates that set

CreateTrainingJob

Starts a model training job. After training completes, Amazon SageMaker saves the resulting

CreateTransformJob

Starts a transform job. A transform job uses a trained model to get inferences on a

CreateTrial

Creates an SageMaker trial. A trial is a set of steps called

CreateTrialComponent

Creates a trial component, which is a stage of a machine learning

CreateUserProfile

Creates a user profile. A user profile represents a single user within a domain, and is

CreateWorkforce

Use this operation to create a workforce. This operation will return an error

CreateWorkteam

Creates a new work team for labeling your data. A work team is defined by one or more

DeleteAction

Deletes an action.

DeleteAlgorithm

Removes the specified algorithm from your account.

DeleteApp

Used to stop and delete an app.

DeleteAppImageConfig

Deletes an AppImageConfig.

DeleteArtifact

Deletes an artifact. Either ArtifactArn or Source must be

DeleteAssociation

Deletes an association.

DeleteCodeRepository

Deletes the specified Git repository from your account.

DeleteContext

Deletes an context.

DeleteDataQualityJobDefinition

Deletes a data quality monitoring job definition.

DeleteDeviceFleet

Deletes a fleet.

DeleteDomain

Used to delete a domain.

DeleteEndpoint

Deletes an endpoint. Amazon SageMaker frees up all of the resources that were deployed when the

DeleteEndpointConfig

Deletes an endpoint configuration. The DeleteEndpointConfig API

DeleteExperiment

Deletes an SageMaker experiment. All trials associated with the experiment must be deleted

DeleteFeatureGroup

Delete the FeatureGroup and any data that was written to the

DeleteFlowDefinition

Deletes the specified flow definition.

DeleteHumanTaskUi

Use this operation to delete a human task user interface (worker task template).

DeleteImage

Deletes a SageMaker image and all versions of the image. The container images aren't

DeleteImageVersion

Deletes a version of a SageMaker image. The container image the version represents isn't

DeleteModel

Deletes a model. The DeleteModel API deletes only the model entry that

DeleteModelBiasJobDefinition

Deletes an Amazon SageMaker model bias job definition.

DeleteModelExplainabilityJobDefinition

Deletes an Amazon SageMaker model explainability job definition.

DeleteModelPackage

Deletes a model package.

DeleteModelPackageGroup

Deletes the specified model group.

DeleteModelPackageGroupPolicy

Deletes a model group resource policy.

DeleteModelQualityJobDefinition

Deletes the secified model quality monitoring job definition.

DeleteMonitoringSchedule

Deletes a monitoring schedule. Also stops the schedule had not already been stopped.

DeleteNotebookInstance

Deletes an Amazon SageMaker notebook instance. Before you can delete a notebook instance, you

DeleteNotebookInstanceLifecycleConfig

Deletes a notebook instance lifecycle configuration.

DeletePipeline

Deletes a pipeline if there are no running instances of the pipeline. To delete a

DeleteProject

Delete the specified project.

DeleteTags

Deletes the specified tags from an Amazon SageMaker resource.

DeleteTrial

Deletes the specified trial. All trial components that make up the trial must be deleted

DeleteTrialComponent

Deletes the specified trial component. A trial component must be disassociated from all

DeleteUserProfile

Deletes a user profile. When a user profile is deleted, the user loses access to their EFS

DeleteWorkforce

Use this operation to delete a workforce.

DeleteWorkteam

Deletes an existing work team. This operation can't be undone.

DeregisterDevices

Deregisters the specified devices. After you deregister a device, you will need to re-register the devices.

DescribeAction

Describes an action.

DescribeAlgorithm

Returns a description of the specified algorithm that is in your account.

DescribeApp

Describes the app.

DescribeAppImageConfig

Describes an AppImageConfig.

DescribeArtifact

Describes an artifact.

DescribeAutoMLJob

Returns information about an Amazon SageMaker AutoML job.

DescribeCodeRepository

Gets details about the specified Git repository.

DescribeCompilationJob

Returns information about a model compilation job.

DescribeContext

Describes a context.

DescribeDataQualityJobDefinition

Gets the details of a data quality monitoring job definition.

DescribeDevice

Describes the device.

DescribeDeviceFleet

A description of the fleet the device belongs to.

DescribeDomain

The description of the domain.

DescribeEdgePackagingJob

A description of edge packaging jobs.

DescribeEndpoint

Returns the description of an endpoint.

DescribeEndpointConfig

Returns the description of an endpoint configuration created using the

DescribeExperiment

Provides a list of an experiment's properties.

DescribeFeatureGroup

Use this operation to describe a FeatureGroup. The response includes

DescribeFlowDefinition

Returns information about the specified flow definition.

DescribeHumanTaskUi

Returns information about the requested human task user interface (worker task template).

DescribeHyperParameterTuningJob

Gets

DescribeImage

Describes a SageMaker image.

DescribeImageVersion

Describes a version of a SageMaker image.

DescribeLabelingJob

Gets information about a labeling job.

DescribeModel

Describes a model that you created using the CreateModel

DescribeModelBiasJobDefinition

Returns a description of a model bias job definition.

DescribeModelExplainabilityJobDefinition

Returns a description of a model explainability job definition.

DescribeModelPackage

Returns a description of the specified model package, which is used to create Amazon SageMaker

DescribeModelPackageGroup

Gets a description for the specified model group.

DescribeModelQualityJobDefinition

Returns a description of a model quality job definition.

DescribeMonitoringSchedule

Describes the schedule for a monitoring job.

DescribeNotebookInstance

Returns information about a notebook instance.

DescribeNotebookInstanceLifecycleConfig

Returns a description of a notebook instance lifecycle configuration.

DescribePipeline

Describes the details of a pipeline.

DescribePipelineDefinitionForExecution

Describes the details of an execution's pipeline definition.

DescribePipelineExecution

Describes the details of a pipeline execution.

DescribeProcessingJob

Returns a description of a processing job.

DescribeProject

Describes the details of a project.

DescribeSubscribedWorkteam

Gets information about a work team provided by a vendor. It returns details about the

DescribeTrainingJob

Returns information about a training job.

DescribeTransformJob

Returns information about a transform job.

DescribeTrial

Provides a list of a trial's properties.

DescribeTrialComponent

Provides a list of a trials component's properties.

DescribeUserProfile

Describes a user profile. For more information, see CreateUserProfile.

DescribeWorkforce

Lists private workforce information, including workforce name, Amazon Resource Name

DescribeWorkteam

Gets information about a specific work team. You can see information such as the

DisableSagemakerServicecatalogPortfolio

Disables using Service Catalog in SageMaker. Service Catalog is used to create

DisassociateTrialComponent

Disassociates a trial component from a trial. This doesn't effect other trials the

EnableSagemakerServicecatalogPortfolio

Enables using Service Catalog in SageMaker. Service Catalog is used to create

GetDeviceFleetReport

Describes a fleet.

GetModelPackageGroupPolicy

Gets a resource policy that manages access for a model group. For information about

GetSagemakerServicecatalogPortfolioStatus

Gets the status of Service Catalog in SageMaker. Service Catalog is used to create

GetSearchSuggestions

An auto-complete API for the search functionality in the Amazon SageMaker console. It returns

ListActions

Lists the actions in your account and their properties.

ListAlgorithms

Lists the machine learning algorithms that have been created.

ListAppImageConfigs

Lists the AppImageConfigs in your account and their properties. The list can be

ListApps

Lists apps.

ListArtifacts

Lists the artifacts in your account and their properties.

ListAssociations

Lists the associations in your account and their properties.

ListAutoMLJobs

Request a list of jobs.

ListCandidatesForAutoMLJob

List the candidates created for the job.

ListCodeRepositories

Gets a list of the Git repositories in your account.

ListCompilationJobs

Lists model compilation jobs that satisfy various filters.

ListContexts

Lists the contexts in your account and their properties.

ListDataQualityJobDefinitions

Lists the data quality job definitions in your account.

ListDeviceFleets

Returns a list of devices in the fleet.

ListDevices

A list of devices.

ListDomains

Lists the domains.

ListEdgePackagingJobs

Returns a list of edge packaging jobs.

ListEndpointConfigs

Lists endpoint configurations.

ListEndpoints

Lists endpoints.

ListExperiments

Lists all the experiments in your account. The list can be filtered to show only

ListFeatureGroups

List FeatureGroups based on given filter and order.

ListFlowDefinitions

Returns information about the flow definitions in your account.

ListHumanTaskUis

Returns information about the human task user interfaces in your account.

ListHyperParameterTuningJobs

Gets a list of HyperParameterTuningJobSummary objects that

ListImageVersions

Lists the versions of a specified image and their properties. The list can be filtered

ListImages

Lists the images in your account and their properties. The list can be filtered by

ListLabelingJobs

Gets a list of labeling jobs.

ListLabelingJobsForWorkteam

Gets a list of labeling jobs assigned to a specified work team.

ListModelBiasJobDefinitions

Lists model bias jobs definitions that satisfy various filters.

ListModelExplainabilityJobDefinitions

Lists model explainability job definitions that satisfy various filters.

ListModelPackageGroups

Gets a list of the model groups in your Amazon Web Services account.

ListModelPackages

Lists the model packages that have been created.

ListModelQualityJobDefinitions

Gets a list of model quality monitoring job definitions in your account.

ListModels

Lists models created with the CreateModel API.

ListMonitoringExecutions

Returns list of all monitoring job executions.

ListMonitoringSchedules

Returns list of all monitoring schedules.

ListNotebookInstanceLifecycleConfigs

Lists notebook instance lifestyle configurations created with the CreateNotebookInstanceLifecycleConfig API.

ListNotebookInstances

Returns a list of the Amazon SageMaker notebook instances in the requester's account in an Amazon Web Services

ListPipelineExecutionSteps

Gets a list of PipeLineExecutionStep objects.

ListPipelineExecutions

Gets a list of the pipeline executions.

ListPipelineParametersForExecution

Gets a list of parameters for a pipeline execution.

ListPipelines

Gets a list of pipelines.

ListProcessingJobs

Lists processing jobs that satisfy various filters.

ListProjects

Gets a list of the projects in an Amazon Web Services account.

ListSubscribedWorkteams

Gets a list of the work teams that you are subscribed to in the Amazon Web Services Marketplace. The

ListTags

Returns the tags for the specified Amazon SageMaker resource.

ListTrainingJobs

Lists training jobs.

ListTrainingJobsForHyperParameterTuningJob

Gets a list of TrainingJobSummary objects that describe the training

ListTransformJobs

Lists transform jobs.

ListTrialComponents

Lists the trial components in your account. You can sort the list by trial component name

ListTrials

Lists the trials in your account. Specify an experiment name to limit the list to the

ListUserProfiles

Lists user profiles.

ListWorkforces

Use this operation to list all private and vendor workforces in an Amazon Web Services Region. Note that you can only

ListWorkteams

Gets a list of private work teams that you have defined in a region. The list may be empty if

PutModelPackageGroupPolicy

Adds a resouce policy to control access to a model group. For information about

RegisterDevices

Register devices.

RenderUiTemplate

Renders the UI template so that you can preview the worker's experience.

Search

Finds Amazon SageMaker resources that match a search query. Matching resources are returned

SendPipelineExecutionStepFailure

Notifies the pipeline that the execution of a callback step failed, along with a

SendPipelineExecutionStepSuccess

Notifies the pipeline that the execution of a callback step succeeded and provides a

StartMonitoringSchedule

Starts a previously stopped monitoring schedule.

StartNotebookInstance

Launches an ML compute instance with the latest version of the libraries and

StartPipelineExecution

Starts a pipeline execution.

StopAutoMLJob

A method for forcing the termination of a running job.

StopCompilationJob

Stops a model compilation job.

StopEdgePackagingJob

Request to stop an edge packaging job.

StopHyperParameterTuningJob

Stops a running hyperparameter tuning job and all running training jobs that the

StopLabelingJob

Stops a running labeling job. A job that is stopped cannot be restarted. Any results

StopMonitoringSchedule

Stops a previously started monitoring schedule.

StopNotebookInstance

Terminates the ML compute instance. Before terminating the instance, Amazon SageMaker

StopPipelineExecution

Stops a pipeline execution.

StopProcessingJob

Stops a processing job.

StopTrainingJob

Stops a training job. To stop a job, Amazon SageMaker sends the algorithm the

StopTransformJob

Stops a transform job.

UpdateAction

Updates an action.

UpdateAppImageConfig

Updates the properties of an AppImageConfig.

UpdateArtifact

Updates an artifact.

UpdateCodeRepository

Updates the specified Git repository with the specified values.

UpdateContext

Updates a context.

UpdateDeviceFleet

Updates a fleet of devices.

UpdateDevices

Updates one or more devices in a fleet.

UpdateDomain

Updates the default settings for new user profiles in the domain.

UpdateEndpoint

Deploys the new EndpointConfig specified in the request, switches to

UpdateEndpointWeightsAndCapacities

Updates variant weight of one or more variants associated with an existing

UpdateExperiment

Adds, updates, or removes the description of an experiment. Updates the display name of an

UpdateImage

Updates the properties of a SageMaker image. To change the image's tags, use the

UpdateModelPackage

Updates a versioned model.

UpdateMonitoringSchedule

Updates a previously created schedule.

UpdateNotebookInstance

Updates a notebook instance. NotebookInstance updates include upgrading or

UpdateNotebookInstanceLifecycleConfig

Updates a notebook instance lifecycle configuration created with the CreateNotebookInstanceLifecycleConfig API.

UpdatePipeline

Updates a pipeline.

UpdatePipelineExecution

Updates a pipeline execution.

UpdateTrainingJob

Update a model training job to request a new Debugger profiling configuration.

UpdateTrial

Updates the display name of a trial.

UpdateTrialComponent

Updates one or more properties of a trial component.

UpdateUserProfile

Updates a user profile.

UpdateWorkforce

Use this operation to update your workforce. You can use this operation to

UpdateWorkteam

Updates an existing work team with new member definitions or description.