Struct aws_sdk_sagemaker::output::describe_compilation_job_output::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for DescribeCompilationJobOutput

Implementations

impl Builder

pub fn compilation_job_name(self, input: impl Into<String>) -> Self

The name of the model compilation job.

pub fn set_compilation_job_name(self, input: Option<String>) -> Self

pub fn compilation_job_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the model compilation job.

pub fn set_compilation_job_arn(self, input: Option<String>) -> Self

pub fn compilation_job_status(self, input: CompilationJobStatus) -> Self

The status of the model compilation job.

pub fn set_compilation_job_status(
self,
input: Option<CompilationJobStatus>
) -> Self

pub fn compilation_start_time(self, input: Instant) -> Self

The time when the model compilation job started the CompilationJob instances.

You are billed for the time between this timestamp and the timestamp in the DescribeCompilationJobResponse$CompilationEndTime field. In Amazon CloudWatch Logs, the start time might be later than this time. That's because it takes time to download the compilation job, which depends on the size of the compilation job container.

pub fn set_compilation_start_time(self, input: Option<Instant>) -> Self

pub fn compilation_end_time(self, input: Instant) -> Self

The time when the model compilation job on a compilation job instance ended. For a successful or stopped job, this is when the job's model artifacts have finished uploading. For a failed job, this is when Amazon SageMaker detected that the job failed.

pub fn set_compilation_end_time(self, input: Option<Instant>) -> Self

pub fn stopping_condition(self, input: StoppingCondition) -> Self

Specifies a limit to how long a model compilation job can run. When the job reaches the time limit, Amazon SageMaker ends the compilation job. Use this API to cap model training costs.

pub fn set_stopping_condition(self, input: Option<StoppingCondition>) -> Self

pub fn inference_image(self, input: impl Into<String>) -> Self

The inference image to use when compiling a model. Specify an image only if the target device is a cloud instance.

pub fn set_inference_image(self, input: Option<String>) -> Self

pub fn creation_time(self, input: Instant) -> Self

The time that the model compilation job was created.

pub fn set_creation_time(self, input: Option<Instant>) -> Self

pub fn last_modified_time(self, input: Instant) -> Self

The time that the status of the model compilation job was last modified.

pub fn set_last_modified_time(self, input: Option<Instant>) -> Self

pub fn failure_reason(self, input: impl Into<String>) -> Self

If a model compilation job failed, the reason it failed.

pub fn set_failure_reason(self, input: Option<String>) -> Self

pub fn model_artifacts(self, input: ModelArtifacts) -> Self

Information about the location in Amazon S3 that has been configured for storing the model artifacts used in the compilation job.

pub fn set_model_artifacts(self, input: Option<ModelArtifacts>) -> Self

pub fn model_digests(self, input: ModelDigests) -> Self

Provides a BLAKE2 hash value that identifies the compiled model artifacts in Amazon S3.

pub fn set_model_digests(self, input: Option<ModelDigests>) -> Self

pub fn role_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of an IAM role that Amazon SageMaker assumes to perform the model compilation job.

pub fn set_role_arn(self, input: Option<String>) -> Self

pub fn input_config(self, input: InputConfig) -> Self

Information about the location in Amazon S3 of the input model artifacts, the name and shape of the expected data inputs, and the framework in which the model was trained.

pub fn set_input_config(self, input: Option<InputConfig>) -> Self

pub fn output_config(self, input: OutputConfig) -> Self

Information about the output location for the compiled model and the target device that the model runs on.

pub fn set_output_config(self, input: Option<OutputConfig>) -> Self

pub fn vpc_config(self, input: NeoVpcConfig) -> Self

A VpcConfig object that specifies the VPC that you want your compilation job to connect to. Control access to your models by configuring the VPC. For more information, see Protect Compilation Jobs by Using an Amazon Virtual Private Cloud.

pub fn set_vpc_config(self, input: Option<NeoVpcConfig>) -> Self

pub fn build(self) -> DescribeCompilationJobOutput

Consumes the builder and constructs a DescribeCompilationJobOutput