Struct aws_sdk_sagemaker::output::describe_hyper_parameter_tuning_job_output::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for DescribeHyperParameterTuningJobOutput

Implementations

impl Builder

pub fn hyper_parameter_tuning_job_name(self, input: impl Into<String>) -> Self

The name of the tuning job.

pub fn set_hyper_parameter_tuning_job_name(self, input: Option<String>) -> Self

pub fn hyper_parameter_tuning_job_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the tuning job.

pub fn set_hyper_parameter_tuning_job_arn(self, input: Option<String>) -> Self

pub fn hyper_parameter_tuning_job_config(
self,
input: HyperParameterTuningJobConfig
) -> Self

The HyperParameterTuningJobConfig object that specifies the configuration of the tuning job.

pub fn set_hyper_parameter_tuning_job_config(
self,
input: Option<HyperParameterTuningJobConfig>
) -> Self

pub fn training_job_definition(
self,
input: HyperParameterTrainingJobDefinition
) -> Self

The HyperParameterTrainingJobDefinition object that specifies the definition of the training jobs that this tuning job launches.

pub fn set_training_job_definition(
self,
input: Option<HyperParameterTrainingJobDefinition>
) -> Self

pub fn training_job_definitions(
self,
input: impl Into<HyperParameterTrainingJobDefinition>
) -> Self

pub fn set_training_job_definitions(
self,
input: Option<Vec<HyperParameterTrainingJobDefinition>>
) -> Self

pub fn hyper_parameter_tuning_job_status(
self,
input: HyperParameterTuningJobStatus
) -> Self

The status of the tuning job: InProgress, Completed, Failed, Stopping, or Stopped.

pub fn set_hyper_parameter_tuning_job_status(
self,
input: Option<HyperParameterTuningJobStatus>
) -> Self

pub fn creation_time(self, input: Instant) -> Self

The date and time that the tuning job started.

pub fn set_creation_time(self, input: Option<Instant>) -> Self

pub fn hyper_parameter_tuning_end_time(self, input: Instant) -> Self

The date and time that the tuning job ended.

pub fn set_hyper_parameter_tuning_end_time(self, input: Option<Instant>) -> Self

pub fn last_modified_time(self, input: Instant) -> Self

The date and time that the status of the tuning job was modified.

pub fn set_last_modified_time(self, input: Option<Instant>) -> Self

pub fn training_job_status_counters(
self,
input: TrainingJobStatusCounters
) -> Self

The TrainingJobStatusCounters object that specifies the number of training jobs, categorized by status, that this tuning job launched.

pub fn set_training_job_status_counters(
self,
input: Option<TrainingJobStatusCounters>
) -> Self

pub fn objective_status_counters(self, input: ObjectiveStatusCounters) -> Self

The ObjectiveStatusCounters object that specifies the number of training jobs, categorized by the status of their final objective metric, that this tuning job launched.

pub fn set_objective_status_counters(
self,
input: Option<ObjectiveStatusCounters>
) -> Self

pub fn best_training_job(self, input: HyperParameterTrainingJobSummary) -> Self

A TrainingJobSummary object that describes the training job that completed with the best current HyperParameterTuningJobObjective.

pub fn set_best_training_job(
self,
input: Option<HyperParameterTrainingJobSummary>
) -> Self

pub fn overall_best_training_job(
self,
input: HyperParameterTrainingJobSummary
) -> Self

If the hyperparameter tuning job is an warm start tuning job with a WarmStartType of IDENTICAL_DATA_AND_ALGORITHM, this is the TrainingJobSummary for the training job with the best objective metric value of all training jobs launched by this tuning job and all parent jobs specified for the warm start tuning job.

pub fn set_overall_best_training_job(
self,
input: Option<HyperParameterTrainingJobSummary>
) -> Self

pub fn warm_start_config(
self,
input: HyperParameterTuningJobWarmStartConfig
) -> Self

The configuration for starting the hyperparameter parameter tuning job using one or more previous tuning jobs as a starting point. The results of previous tuning jobs are used to inform which combinations of hyperparameters to search over in the new tuning job.

pub fn set_warm_start_config(
self,
input: Option<HyperParameterTuningJobWarmStartConfig>
) -> Self

pub fn failure_reason(self, input: impl Into<String>) -> Self

If the tuning job failed, the reason it failed.

pub fn set_failure_reason(self, input: Option<String>) -> Self

pub fn build(self) -> DescribeHyperParameterTuningJobOutput

Consumes the builder and constructs a DescribeHyperParameterTuningJobOutput