Struct aws_sdk_sagemaker::output::DescribeProcessingJobOutput

#[non_exhaustive]
pub struct DescribeProcessingJobOutput {

    pub processing_inputs: Option<Vec<ProcessingInput>>,
    pub processing_output_config: Option<ProcessingOutputConfig>,
    pub processing_job_name: Option<String>,
    pub processing_resources: Option<ProcessingResources>,
    pub stopping_condition: Option<ProcessingStoppingCondition>,
    pub app_specification: Option<AppSpecification>,
    pub environment: Option<HashMap<String, String>>,
    pub network_config: Option<NetworkConfig>,
    pub role_arn: Option<String>,
    pub experiment_config: Option<ExperimentConfig>,
    pub processing_job_arn: Option<String>,
    pub processing_job_status: Option<ProcessingJobStatus>,
    pub exit_message: Option<String>,
    pub failure_reason: Option<String>,
    pub processing_end_time: Option<Instant>,
    pub processing_start_time: Option<Instant>,
    pub last_modified_time: Option<Instant>,
    pub creation_time: Option<Instant>,
    pub monitoring_schedule_arn: Option<String>,
    pub auto_ml_job_arn: Option<String>,
    pub training_job_arn: Option<String>,

}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

processing_inputs: Option<Vec<ProcessingInput>>

The inputs for a processing job.

processing_output_config: Option<ProcessingOutputConfig>

Output configuration for the processing job.

processing_job_name: Option<String>

The name of the processing job. The name must be unique within an Amazon Web Services Region in the Amazon Web Services account.

processing_resources: Option<ProcessingResources>

Identifies the resources, ML compute instances, and ML storage volumes to deploy for a processing job. In distributed training, you specify more than one instance.

stopping_condition: Option<ProcessingStoppingCondition>

The time limit for how long the processing job is allowed to run.

app_specification: Option<AppSpecification>

Configures the processing job to run a specified container image.

environment: Option<HashMap<String, String>>

The environment variables set in the Docker container.

network_config: Option<NetworkConfig>

Networking options for a processing job.

role_arn: Option<String>

The Amazon Resource Name (ARN) of an IAM role that Amazon SageMaker can assume to perform tasks on your behalf.

experiment_config: Option<ExperimentConfig>

The configuration information used to create an experiment.

processing_job_arn: Option<String>

The Amazon Resource Name (ARN) of the processing job.

processing_job_status: Option<ProcessingJobStatus>

Provides the status of a processing job.

exit_message: Option<String>

An optional string, up to one KB in size, that contains metadata from the processing container when the processing job exits.

failure_reason: Option<String>

A string, up to one KB in size, that contains the reason a processing job failed, if it failed.

processing_end_time: Option<Instant>

The time at which the processing job completed.

processing_start_time: Option<Instant>

The time at which the processing job started.

last_modified_time: Option<Instant>

The time at which the processing job was last modified.

creation_time: Option<Instant>

The time at which the processing job was created.

monitoring_schedule_arn: Option<String>

The ARN of a monitoring schedule for an endpoint associated with this processing job.

auto_ml_job_arn: Option<String>

The ARN of an AutoML job associated with this processing job.

training_job_arn: Option<String>

The ARN of a training job associated with this processing job.

Implementations

impl DescribeProcessingJobOutput

pub fn builder() -> Builder

Creates a new builder-style object to manufacture DescribeProcessingJobOutput