Struct aws_sdk_sagemaker::output::DescribeTransformJobOutput

#[non_exhaustive]
pub struct DescribeTransformJobOutput {

    pub transform_job_name: Option<String>,
    pub transform_job_arn: Option<String>,
    pub transform_job_status: Option<TransformJobStatus>,
    pub failure_reason: Option<String>,
    pub model_name: Option<String>,
    pub max_concurrent_transforms: Option<i32>,
    pub model_client_config: Option<ModelClientConfig>,
    pub max_payload_in_mb: Option<i32>,
    pub batch_strategy: Option<BatchStrategy>,
    pub environment: Option<HashMap<String, String>>,
    pub transform_input: Option<TransformInput>,
    pub transform_output: Option<TransformOutput>,
    pub transform_resources: Option<TransformResources>,
    pub creation_time: Option<Instant>,
    pub transform_start_time: Option<Instant>,
    pub transform_end_time: Option<Instant>,
    pub labeling_job_arn: Option<String>,
    pub auto_ml_job_arn: Option<String>,
    pub data_processing: Option<DataProcessing>,
    pub experiment_config: Option<ExperimentConfig>,

}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

transform_job_name: Option<String>

The name of the transform job.

transform_job_arn: Option<String>

The Amazon Resource Name (ARN) of the transform job.

transform_job_status: Option<TransformJobStatus>

The status of the transform job. If the transform job failed, the reason is returned in the FailureReason field.

failure_reason: Option<String>

If the transform job failed, FailureReason describes why it failed. A transform job creates a log file, which includes error messages, and stores it as an Amazon S3 object. For more information, see Log Amazon SageMaker Events with Amazon CloudWatch.

model_name: Option<String>

The name of the model used in the transform job.

max_concurrent_transforms: Option<i32>

The maximum number of parallel requests on each instance node that can be launched in a transform job. The default value is 1.

model_client_config: Option<ModelClientConfig>

The timeout and maximum number of retries for processing a transform job invocation.

max_payload_in_mb: Option<i32>

The maximum payload size, in MB, used in the transform job.

batch_strategy: Option<BatchStrategy>

Specifies the number of records to include in a mini-batch for an HTTP inference request. A record is a single unit of input data that inference can be made on. For example, a single line in a CSV file is a record.

To enable the batch strategy, you must set SplitType to Line, RecordIO, or TFRecord.

environment: Option<HashMap<String, String>>

The environment variables to set in the Docker container. We support up to 16 key and values entries in the map.

transform_input: Option<TransformInput>

Describes the dataset to be transformed and the Amazon S3 location where it is stored.

transform_output: Option<TransformOutput>

Identifies the Amazon S3 location where you want Amazon SageMaker to save the results from the transform job.

transform_resources: Option<TransformResources>

Describes the resources, including ML instance types and ML instance count, to use for the transform job.

creation_time: Option<Instant>

A timestamp that shows when the transform Job was created.

transform_start_time: Option<Instant>

Indicates when the transform job starts on ML instances. You are billed for the time interval between this time and the value of TransformEndTime.

transform_end_time: Option<Instant>

Indicates when the transform job has been completed, or has stopped or failed. You are billed for the time interval between this time and the value of TransformStartTime.

labeling_job_arn: Option<String>

The Amazon Resource Name (ARN) of the Amazon SageMaker Ground Truth labeling job that created the transform or training job.

auto_ml_job_arn: Option<String>

The Amazon Resource Name (ARN) of the AutoML transform job.

data_processing: Option<DataProcessing>

The data structure used to specify the data to be used for inference in a batch transform job and to associate the data that is relevant to the prediction results in the output. The input filter provided allows you to exclude input data that is not needed for inference in a batch transform job. The output filter provided allows you to include input data relevant to interpreting the predictions in the output from the job. For more information, see Associate Prediction Results with their Corresponding Input Records.

experiment_config: Option<ExperimentConfig>

Associates a SageMaker job as a trial component with an experiment and trial. Specified when you call the following APIs:

CreateProcessingJob

CreateTrainingJob

CreateTransformJob

Implementations

impl DescribeTransformJobOutput

pub fn builder() -> Builder

Creates a new builder-style object to manufacture DescribeTransformJobOutput